% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gisco-attributions.R
\encoding{UTF-8}
\name{gisco_attributions}
\alias{gisco_attributions}
\title{Attribution for administrative and statistical GISCO data}
\usage{
gisco_attributions(lang = "en", copyright = FALSE)
}
\arguments{
\item{lang}{A character string with the language (two-letter ISO code). See
\link[countrycode:codelist]{countrycode::codelist} and \strong{Details}.}

\item{copyright}{A logical value indicating whether to display the
copyright notice on the console.}
}
\value{
A string with the attribution to be used.
}
\description{
Get the attribution text to use for administrative and statistical data
downloaded from GISCO. See the \strong{Copyright} section.

For other datasets you may follow the \href{https://ec.europa.eu/eurostat/web/main/help/copyright-notice}{Eurostat general copyright notice and license policy}.
}
\details{
Supported languages are:
\itemize{
\item \code{"en"}: English.
\item \code{"da"}: Danish.
\item \code{"de"}: German.
\item \code{"es"}: Spanish.
\item \code{"fi"}: Finnish.
\item \code{"fr"}: French.
\item \code{"no"}: Norwegian.
\item \code{"sv"}: Swedish.
}

Please consider
\href{https://github.com/rOpenGov/giscoR/issues}{contributing} if you spot a
mistake or want to add a new language.
}
\section{Copyright}{
The provisions described in this section apply to administrative and
statistical data provided by the following functions:

\strong{Administrative units}
\itemize{
\item \code{\link[=gisco_get_communes]{gisco_get_communes()}}.
\item \code{\link[=gisco_get_countries]{gisco_get_countries()}}.
}

\strong{Statistical units}
\itemize{
\item \code{\link[=gisco_get_census]{gisco_get_census()}}.
\item \code{\link[=gisco_get_coastal_lines]{gisco_get_coastal_lines()}}.
\item \code{\link[=gisco_get_lau]{gisco_get_lau()}}.
\item \code{\link[=gisco_get_nuts]{gisco_get_nuts()}}.
\item \code{\link[=gisco_get_urban_audit]{gisco_get_urban_audit()}}.
}

Postal code data use a separate CC-BY-SA-4.0 attribution. See
\code{\link[=gisco_get_postal_codes]{gisco_get_postal_codes()}}.
\subsection{Copyright notice}{

When data downloaded from GISCO is used in any printed or electronic
publication, in addition to any other provisions applicable to the whole
Eurostat website, the data source must be acknowledged in the legend of
the map and on the introductory page of the publication with the following
copyright notice:
\itemize{
\item EN: © EuroGeographics for the administrative boundaries.
\item FR: © EuroGeographics pour les limites administratives.
\item DE: © EuroGeographics bezüglich der Verwaltungsgrenzen.
}

For publications in languages other than English, French or German,
the translation of the copyright notice in the language of the publication
shall be used.

If you intend to use the data commercially, please contact EuroGeographics
for information about their license agreements.
}
}

\examples{
gisco_attributions()

gisco_attributions(lang = "es", copyright = TRUE)

gisco_attributions(lang = "XXX")

# Get the list of codes from countrycode.
if (requireNamespace("dplyr", quietly = TRUE)) {
  library(dplyr)

  countrycode::codelist |>
    select(country.name.en, iso2c)
}

}
\concept{attribution}
