% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_downloads.R
\name{gisco_check_access}
\alias{gisco_check_access}
\title{Check access to GISCO API}
\usage{
gisco_check_access()
}
\value{
a logical.
}
\description{
Check if R has access to resources at
\url{https://gisco-services.ec.europa.eu/distribution/v2/}.
}
\examples{

gisco_check_access()
}
\concept{helper}
