% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gisco-cache.R
\encoding{UTF-8}
\name{gisco_clear_cache}
\alias{gisco_clear_cache}
\title{Clear your \CRANpkg{giscoR} cache directory}
\usage{
gisco_clear_cache(config = FALSE, cached_data = TRUE, verbose = FALSE)
}
\arguments{
\item{config}{If \code{TRUE}, delete the configuration folder of
\CRANpkg{giscoR}.}

\item{cached_data}{If \code{TRUE}, delete your \code{cache_dir} and all its contents.}

\item{verbose}{A logical value indicating whether to display informational
messages.}
}
\value{
Invisible. Called for its side effects.
}
\description{
\strong{Use this function with caution}. It clears your cached data and
configuration, specifically:
\itemize{
\item Deletes the \CRANpkg{giscoR} config directory
(\code{tools::R_user_dir("giscoR", "config")}).
\item Deletes the \code{cache_dir} directory.
\item Deletes the value stored in \code{Sys.getenv("GISCO_CACHE_DIR")}.
}
}
\details{
Fully resets your cache state as if you had never installed or used
\CRANpkg{giscoR}.
}
\examples{

# Do not run this. It modifies your current state.
\dontrun{
my_cache <- gisco_detect_cache_dir()

# Set an example cache.
ex <- file.path(tempdir(), "example", "cache")
gisco_set_cache_dir(ex, verbose = FALSE)

# Restore the initial cache.
gisco_clear_cache(verbose = TRUE)

gisco_set_cache_dir(my_cache)
identical(my_cache, gisco_detect_cache_dir())
}
}
\seealso{
\code{\link[tools:R_user_dir]{tools::R_user_dir()}}

Cache management utilities:
\code{\link[=gisco_set_cache_dir]{gisco_set_cache_dir()}}
}
\concept{cache_utils}
