% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{gisco_coastal_lines}
\alias{gisco_coastal_lines}
\title{Coastal lines 2016 \code{\link[sf:st_sf]{sf}} object}
\format{
An \code{\link[sf:st_sf]{sf}} object with \code{POLYGON} geometries, resolution:
1:20 million and \href{https://epsg.io/4326}{EPSG:4326}.
}
\source{
\href{https://gisco-services.ec.europa.eu/distribution/v2/coas/gpkg/}{COAS_RG_20M_2016_4326.gpkg} file.
}
\description{
This object contains the coastal lines of the world.
}
\section{Copyright}{
See the GISCO statistical unit copyright provisions:
\url{https://ec.europa.eu/eurostat/web/gisco/geodata/statistical-units}.
}

\section{Note}{
Check the download and usage provisions in \code{\link[=gisco_attributions]{gisco_attributions()}}.
}

\examples{
library(sf)
data("gisco_coastal_lines")
gisco_coastal_lines

}
\seealso{
\code{\link[=gisco_get_coastal_lines]{gisco_get_coastal_lines()}}

Included datasets:
\code{\link{gisco_countries_2024}},
\code{\link{gisco_countrycode}},
\code{\link{gisco_db}},
\code{\link{gisco_nuts_2024}}
}
\concept{datasets}
\keyword{datasets}
