% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gisco_coastallines_60M_2016}
\alias{gisco_coastallines_60M_2016}
\title{World coastal lines \code{LINESTRING} object}
\format{
A \code{LINESTRING} data frame (resolution: 1:60million, EPSG:4326) object with 8 variables:
\describe{
  \item{EFTA_FLAG}{Coast belonging to EFTA countries}
  \item{OTHR_FLAG}{Coast belonging to other countries}
  \item{EU_FLAG}{Coast belonging to EU countries}
  \item{COAS_FLAG}{Coast flag}
  \item{CNTR_BN_ID}{CNTR_BN_ID}
  \item{CC_FLAG}{Coast belonging to EU candidate countries}
  \item{FID}{FID}
  \item{geometry}{geometry field}
}
}
\source{
\url{https://gisco-services.ec.europa.eu/distribution/v2/countries/geojson/CNTR_BN_60M_2016_4326_COASTL.geojson}.
}
\description{
A \code{sf} object including the coast lines
as provided by GISCO (2016 version).
}
