% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{gisco_countries_2024}
\alias{gisco_countries_2024}
\title{Countries 2024 \code{\link[sf:st_sf]{sf}} object}
\format{
An \code{\link[sf:st_sf]{sf}} object with \code{MULTIPOLYGON} geometries, resolution:
1:20 million and \href{https://epsg.io/4326}{EPSG:4326}. It has
263 rows and 12 variables:
\describe{
\item{\code{CNTR_ID}}{Country ID from Eurostat.}
\item{\code{CNTR_NAME}}{Official country name in the local language.}
\item{\code{NAME_ENGL}}{Country name in English.}
\item{\code{NAME_FREN}}{Country name in French.}
\item{\code{ISO3_CODE}}{ISO 3166-1 alpha-3 code of each country from
GISCO.}
\item{\code{SVRG_UN}}{Sovereign status according to the United Nations.}
\item{\code{CAPT}}{Capital city.}
\item{\code{EU_STAT}}{European Union member.}
\item{\code{EFTA_STAT}}{EFTA member.}
\item{\code{CC_STAT}}{EU candidate member.}
\item{\code{NAME_GERM}}{Country name in German.}
\item{\code{geometry}}{Geometry field.}
}
}
\source{
\href{https://gisco-services.ec.europa.eu/distribution/v2/countries/gpkg/}{CNTR_RG_20M_2024_4326.gpkg} file.
}
\description{
This object contains world administrative boundaries at the country level.
}
\section{Copyright}{
See the GISCO administrative unit copyright provisions:
\url{https://ec.europa.eu/eurostat/web/gisco/geodata/administrative-units}.
}

\section{Note}{
Check the download and usage provisions in \code{\link[=gisco_attributions]{gisco_attributions()}}.
}

\examples{

data("gisco_countries_2024")
head(gisco_countries_2024)

}
\seealso{
\code{\link[=gisco_get_countries]{gisco_get_countries()}}

Included datasets:
\code{\link{gisco_coastal_lines}},
\code{\link{gisco_countrycode}},
\code{\link{gisco_db}},
\code{\link{gisco_nuts_2024}}
}
\concept{datasets}
\keyword{datasets}
