% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gisco_countries_20M_2016}
\alias{gisco_countries_20M_2016}
\title{World countries \code{POLYGON} object}
\format{
A \code{MULTIPOLYGON} data frame (resolution: 1:20million, EPSG:4326) object with 257 rows and 7 variables:
\describe{
  \item{id}{row ID}
  \item{CNTR_NAME}{Official country name on local language}
  \item{ISO3_CODE}{\href{https://en.wikipedia.org/wiki/ISO_3166-1_alpha-3}{ISO 3166-1 alpha-3 code} of each country, as provided by GISCO}
  \item{CNTR_ID}{Country ID}
  \item{NAME_ENGL}{Country name in English}
  \item{FID}{FID}
  \item{geometry}{geometry field}
}
}
\source{
\url{https://gisco-services.ec.europa.eu/distribution/v2/countries/geojson/CNTR_RG_20M_2016_4326.geojson}.
}
\description{
A \code{sf} object including all
countries as provided by GISCO (2016 version).
}
\seealso{
\link{gisco_get_countries}
}
