% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{gisco_countrycode}
\alias{gisco_countrycode}
\title{Database with different country code schemes and world regions}
\format{
A data frame with
249 rows and 13
variables:
\describe{
\item{\code{ISO3_CODE}}{ISO 3166-1 alpha-3 code of each country.}
\item{\code{CNTR_CODE}}{Eurostat country code.}
\item{\code{iso2c}}{ISO 3166-1 alpha-2 code of each country.}
\item{\code{iso.name.en}}{ISO English short name.}
\item{\code{cldr.short.en}}{English short name as provided by the Unicode
Common Locale Data Repository.}
\item{\code{continent}}{Continent from the World Bank.}
\item{\code{un.region.code}}{UN M49 numeric region code.}
\item{\code{un.region.name}}{UN M49 region name.}
\item{\code{un.regionintermediate.code}}{UN M49 numeric intermediate region
code.}
\item{\code{un.regionintermediate.name}}{UN M49 intermediate region name.}
\item{\code{un.regionsub.code}}{UN M49 numeric subregion code.}
\item{\code{un.regionsub.name}}{UN M49 subregion name.}
\item{\code{eu}}{Logical value indicating whether the country belongs to the
European Union.}
}
}
\source{
\link[countrycode:codelist]{countrycode::codelist} \strong{v1.6.1}.
}
\description{
A \link[tibble:tbl_df]{tibble} containing conversions between country code
schemes (Eurostat, ISO 3166-1 alpha-2 and ISO 3166-1 alpha-3) and
geographic regions from the World Bank and the UN
(\href{https://unstats.un.org/unsd/methodology/m49/}{M49 Standard}). This
database was extracted from the \CRANpkg{countrycode} package.
}
\section{World regions}{
Regions follow the UN geographic regions
(see \url{https://unstats.un.org/unsd/methodology/m49/}).
Under this scheme Cyprus is assigned to Asia.
}

\examples{
\dontshow{if (requireNamespace("dplyr", quietly = TRUE)) withAutoprint(\{ # examplesIf}
data("gisco_countrycode")
dplyr::glimpse(gisco_countrycode)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=gisco_get_countries]{gisco_get_countries()}}, \link[countrycode:codelist]{countrycode::codelist}.

See also \href{https://cldr.unicode.org/translation/displaynames/countryregion-territory-names}{Unicode Common Locale Data Repository}.

Included datasets:
\code{\link{gisco_coastal_lines}},
\code{\link{gisco_countries_2024}},
\code{\link{gisco_db}},
\code{\link{gisco_nuts_2024}}
}
\concept{datasets}
\keyword{datasets}
