% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{gisco_db}
\alias{gisco_db}
\title{Cached GISCO database}
\format{
A \link[tibble:tbl_df]{tibble} with
11,173 rows.
}
\source{
GISCO geodata distribution \code{datasets.json}.
}
\description{
Database with the list of files in the GISCO geodata distribution as of
2026-09-30.
}
\details{
This database is used to redirect the corresponding functions to the
correct API endpoints.

This version of the database is used if an update fails.
Please use \code{\link[=gisco_get_cached_db]{gisco_get_cached_db()}} with \code{update_cache = TRUE} to update the
corresponding API endpoints.
}
\examples{
\dontshow{if (requireNamespace("dplyr", quietly = TRUE)) withAutoprint(\{ # examplesIf}
data("gisco_db")
gisco_db |>
  dplyr::glimpse()
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=gisco_get_metadata]{gisco_get_metadata()}} and \code{\link[=gisco_get_cached_db]{gisco_get_cached_db()}}.

Included datasets:
\code{\link{gisco_coastal_lines}},
\code{\link{gisco_countries_2024}},
\code{\link{gisco_countrycode}},
\code{\link{gisco_nuts_2024}}

GISCO database and metadata:
\code{\link[=gisco_get_cached_db]{gisco_get_cached_db()}},
\code{\link[=gisco_get_metadata]{gisco_get_metadata()}}
}
\concept{database}
\concept{datasets}
\keyword{datasets}
