% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gisco_db_units}
\alias{gisco_db_units}
\title{Cached GISCO database for units}
\format{
A data frame
}
\source{
GISCO API \code{datasets.json}.
}
\description{
Database with the list of files that the package can load in units.
}
\details{
This data frame is used to check the validity of the API calls.
}
\examples{

data("gisco_db_units")
}
\seealso{
Other dataset: 
\code{\link{gisco_coastallines}},
\code{\link{gisco_countries}},
\code{\link{gisco_countries_2024}},
\code{\link{gisco_countrycode}},
\code{\link{gisco_db}},
\code{\link{gisco_nuts}},
\code{\link{gisco_nuts_2024}}

Other database: 
\code{\link{gisco_db}},
\code{\link{gisco_get_metadata}()}
}
\concept{database}
\concept{dataset}
\keyword{internal}
