% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gisco_cache.R
\name{gisco_detect_cache_dir}
\alias{gisco_detect_cache_dir}
\title{Helper function to detect current \code{cache_dir}}
\usage{
gisco_detect_cache_dir(...)
}
\arguments{
\item{...}{Ignored}
}
\value{
The path to the \code{cache_dir} used in this session
}
\description{
Detect the path to your current \code{cache_dir}. See \code{\link[=gisco_set_cache_dir]{gisco_set_cache_dir()}}
}
\examples{

gisco_detect_cache_dir()

}
\keyword{internal}
