% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gisco-get-cached-db.R
\encoding{UTF-8}
\name{gisco_get_cached_db}
\alias{gisco_get_cached_db}
\title{Retrieve and update the GISCO database used by \CRANpkg{giscoR}}
\source{
\url{https://gisco-services.ec.europa.eu/distribution/v2/}.
}
\usage{
gisco_get_cached_db(update_cache = FALSE)
}
\arguments{
\item{update_cache}{A logical value. If \code{TRUE}, rebuild the cached database
with the most recent information from the GISCO geodata distribution.}
}
\value{
A \link[tibble:tbl_df]{tibble}.
}
\description{
Returns or optionally updates the cached database with endpoints from the
GISCO geodata distribution.
}
\details{
The cached database is stored in the \CRANpkg{giscoR} cache path. See
\code{\link[=gisco_set_cache_dir]{gisco_set_cache_dir()}} for details. The cached database is used in
subsequent \R sessions.

On new GISCO data releases, you can access the updated data by refreshing
the cached database without waiting for a new version of
\CRANpkg{giscoR}.

A static database \link{gisco_db} is shipped with the package. This database is
used if there is any problem during the update.
}
\examples{
\dontshow{if (gisco_check_access()) withAutoprint(\{ # examplesIf}

gisco_get_cached_db() |>
  dplyr::glimpse()
\dontshow{\}) # examplesIf}
}
\seealso{
Database utilities:
\code{\link{gisco_db}},
\code{\link[=gisco_get_metadata]{gisco_get_metadata()}}
}
\concept{database}
