% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gisco-get-census.R
\encoding{UTF-8}
\name{gisco_get_census}
\alias{gisco_get_census}
\title{Census dataset}
\source{
\url{https://ec.europa.eu/eurostat/web/gisco/geodata/statistical-units/census}.
}
\usage{
gisco_get_census(
  year = 2011,
  cache_dir = NULL,
  update_cache = FALSE,
  verbose = FALSE,
  spatialtype = c("RG", "PT")
)
}
\arguments{
\item{year}{A character string or numeric value with the release year of the
file.
Currently only \code{"2011"} is provided.}

\item{cache_dir}{A character string with a path to a cache directory. See
the \strong{Caching strategies} section in \code{\link[=gisco_set_cache_dir]{gisco_set_cache_dir()}}.}

\item{update_cache}{A logical value indicating whether to refresh the
cached file. Defaults to \code{FALSE}. When set to \code{TRUE}, it forces a new
download.}

\item{verbose}{A logical value indicating whether to display informational
messages.}

\item{spatialtype}{The type of geometry to return:
\itemize{
\item \code{"PT"}: Points - \code{POINT} object.
\item \code{"RG"}: Regions - \code{MULTIPOLYGON/POLYGON} object.
}}
}
\value{
An \code{\link[sf:st_sf]{sf}} object.
}
\description{
This dataset provides 2011 census units as region or point geometries.
}
\section{Copyright}{
See the GISCO statistical unit copyright provisions:
\url{https://ec.europa.eu/eurostat/web/gisco/geodata/statistical-units}.
}

\section{Note}{
Check the download and usage provisions in \code{\link[=gisco_attributions]{gisco_attributions()}}.
}

\examples{
\dontshow{if (gisco_check_access()) withAutoprint(\{ # examplesIf}
\donttest{
library(sf)

pts <- gisco_get_census(spatialtype = "PT")

pts
}
\dontshow{\}) # examplesIf}
}
\seealso{
See \code{\link[=gisco_id_api_census_grid]{gisco_id_api_census_grid()}} to download via the GISCO ID service API.

Statistical unit datasets:
\code{\link[=gisco_get_coastal_lines]{gisco_get_coastal_lines()}},
\code{\link[=gisco_get_lau]{gisco_get_lau()}},
\code{\link[=gisco_get_nuts]{gisco_get_nuts()}},
\code{\link[=gisco_get_urban_audit]{gisco_get_urban_audit()}}
}
\concept{stats}
