% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gisco-latest.R
\name{gisco_get_db_years}
\alias{gisco_get_db_years}
\title{Metadata utils (TODO)}
\usage{
gisco_get_db_years(
  id_giscoR = c("coastallines", "communes", "countries", "lau", "nuts", "postalcodes",
    "urban_audit")
)
}
\arguments{
\item{id_giscoR}{Unit types available. Accepted values are
\code{"coastallines"}, \code{"communes"}, \code{"countries"}, \code{"lau"}, \code{"nuts"}, \code{"postalcodes"} or \code{"urban_audit"}.}
}
\description{
Get the latest information available in the current database.
}
\seealso{
Other databases: 
\code{\link{gisco_db}},
\code{\link{gisco_get_latest_db}()}
}
\concept{databases}
