% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gisco-databases.R
\name{gisco_get_latest_db}
\alias{gisco_get_latest_db}
\title{TODO}
\usage{
gisco_get_latest_db(update_cache = FALSE)
}
\arguments{
\item{update_cache}{A logical whether to update cache. Default is \code{FALSE}.
When set to \code{TRUE} it would force a fresh download of the source
\code{.geojson} file.}
}
\description{
TODO Description
}
\seealso{
Other databases: 
\code{\link{gisco_db}},
\code{\link{gisco_get_db_years}()}
}
\concept{databases}
