% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gisco-get-lau.R
\encoding{UTF-8}
\name{gisco_get_lau}
\alias{gisco_get_lau}
\title{Local Administrative Units (LAU) dataset}
\source{
GISCO Local Administrative Units distribution API:
\url{https://gisco-services.ec.europa.eu/distribution/v2/lau/}.
}
\usage{
gisco_get_lau(
  year = 2024,
  epsg = 4326,
  cache = deprecated(),
  update_cache = FALSE,
  cache_dir = NULL,
  verbose = FALSE,
  country = NULL,
  gisco_id = NULL,
  ext = "gpkg"
)
}
\arguments{
\item{year}{A character string or numeric value with the release year of the
file. One of
\Sexpr[stage=render,results=rd]{giscoR:::db_values("lau",
"year",TRUE)}.}

\item{epsg}{A character string or numeric value with the coordinate
reference system as a 4-digit \href{https://epsg.io/}{EPSG code}. One of:
\itemize{
\item \code{"4326"}: \href{https://epsg.io/4326}{WGS84}.
\item \code{"3035"}: \href{https://epsg.io/3035}{ETRS89 / ETRS-LAEA}.
\item \code{"3857"}: \href{https://epsg.io/3857}{Pseudo-Mercator}.
}}

\item{cache}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}. Always caches the result
due to its size. See the \strong{Caching strategies} section in
\code{\link[=gisco_set_cache_dir]{gisco_set_cache_dir()}}.}

\item{update_cache}{A logical value indicating whether to refresh the
cached file. Defaults to \code{FALSE}. When set to \code{TRUE}, it forces a new
download.}

\item{cache_dir}{A character string with a path to a cache directory. See
the \strong{Caching strategies} section in \code{\link[=gisco_set_cache_dir]{gisco_set_cache_dir()}}.}

\item{verbose}{A logical value indicating whether to display informational
messages.}

\item{country}{A character vector of country codes. It can be either a
vector of country names, a vector of ISO 3166-1 alpha-3 country codes or a
vector of Eurostat country codes. See also \code{\link[countrycode:countrycode]{countrycode::countrycode()}}.}

\item{gisco_id}{An optional character vector of \code{GISCO_ID} LAU values.}

\item{ext}{A character string with the extension of the file (default
\code{"gpkg"}). One of
\Sexpr[stage=render,results=rd]{giscoR:::db_values("lau",
"ext",TRUE)}.}
}
\value{
An \code{\link[sf:st_sf]{sf}} object.
}
\description{
Local Administrative Units are the building blocks of NUTS and other
statistical regions. They comprise municipalities and communes in the
European Statistical System. See \code{\link[=gisco_get_communes]{gisco_get_communes()}} for the separate
commune dataset.
}
\details{
The Nomenclature of Territorial Units for Statistics (NUTS) and the LAU
nomenclature are hierarchical classifications of statistical regions that
together subdivide the EU economic territory into statistical regions and
local units, moving from larger to smaller territorial units: NUTS 1, 2 and
3 and LAU.

The LAU classification is not covered by any legislative act. Geographical
extent covers the European Union, EFTA countries and candidate countries.
The scale of the dataset is 1:100 000.

The data contain the National Statistical Agency LAU code, which can be
joined to LAU lists. They also contain a \code{GISCO_ID} field, which is a
unique identifier consisting of the country code and LAU code.

Total resident population figures (31 December) have also been added in
some versions based on the associated LAU lists.
}
\section{Copyright}{
See the GISCO statistical unit copyright provisions:
\url{https://ec.europa.eu/eurostat/web/gisco/geodata/statistical-units}.
}

\section{Note}{
Check the download and usage provisions in \code{\link[=gisco_attributions]{gisco_attributions()}}.
}

\examples{
\dontshow{if (gisco_check_access()) withAutoprint(\{ # examplesIf}
\dontrun{

lu_lau <- gisco_get_lau(year = 2024, country = "Luxembourg")

if (!is.null(lu_lau) && requireNamespace("ggplot2", quietly = TRUE)) {
  library(ggplot2)

  ggplot(lu_lau) +
    geom_sf(aes(fill = POP_DENS_2024)) +
    labs(
      title = "Population density in Luxembourg",
      subtitle = "Year 2024",
      caption = gisco_attributions()
    ) +
    scale_fill_viridis_b(
      option = "cividis",
      label = \(x) prettyNum(x, big.mark = ",")
    ) +
    theme_void() +
    labs(fill = "pop/km2")
}
}
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=gisco_get_communes]{gisco_get_communes()}}.

See \code{\link[=gisco_bulk_download]{gisco_bulk_download()}} to perform a bulk download of datasets.

See \code{\link[=gisco_id_api_lau]{gisco_id_api_lau()}} to download via the GISCO ID service API.

Statistical unit datasets:
\code{\link[=gisco_get_census]{gisco_get_census()}},
\code{\link[=gisco_get_coastal_lines]{gisco_get_coastal_lines()}},
\code{\link[=gisco_get_nuts]{gisco_get_nuts()}},
\code{\link[=gisco_get_urban_audit]{gisco_get_urban_audit()}}
}
\concept{stats}
