% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gisco-get-metadata.R
\encoding{UTF-8}
\name{gisco_get_metadata}
\alias{gisco_get_metadata}
\title{Get metadata}
\source{
\url{https://gisco-services.ec.europa.eu/distribution/v2/}.
}
\usage{
gisco_get_metadata(
  id = c("nuts", "countries", "urban_audit"),
  year = 2024,
  verbose = FALSE
)
}
\arguments{
\item{id}{A character string with the unit type to download. Accepted values
are \code{"nuts"}, \code{"countries"} or \code{"urban_audit"}.}

\item{year}{A character string or numeric value with the release year of the
metadata.}

\item{verbose}{A logical value indicating whether to display informational
messages.}
}
\value{
A \link[tibble:tbl_df]{tibble}.
}
\description{
Get a table with names and IDs of administrative and statistical units.
}
\examples{
\dontshow{if (gisco_check_access()) withAutoprint(\{ # examplesIf}
cities <- gisco_get_metadata(id = "urban_audit", year = 2020)

cities
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=gisco_bulk_download]{gisco_bulk_download()}} for bulk downloads and
\code{\link{gisco_get_unit}} for single-unit downloads.

GISCO database and metadata:
\code{\link{gisco_db}},
\code{\link[=gisco_get_cached_db]{gisco_get_cached_db()}}
}
\concept{database}
