% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gisco-get-nuts.R
\encoding{UTF-8}
\name{gisco_get_nuts}
\alias{gisco_get_nuts}
\title{NUTS statistical units dataset}
\source{
GISCO NUTS distribution API:
\url{https://gisco-services.ec.europa.eu/distribution/v2/nuts/}.
}
\usage{
gisco_get_nuts(
  year = 2024,
  epsg = 4326,
  cache = TRUE,
  update_cache = FALSE,
  cache_dir = NULL,
  verbose = FALSE,
  resolution = 20,
  spatialtype = "RG",
  country = NULL,
  nuts_id = NULL,
  nuts_level = c("all", "0", "1", "2", "3"),
  ext = "gpkg"
)
}
\arguments{
\item{year}{A character string or numeric value with the release year of the
file. One of
\Sexpr[stage=render,results=rd]{giscoR:::db_values("nuts",
"year",TRUE)}.}

\item{epsg}{A character string or numeric value with the coordinate
reference system as a 4-digit \href{https://epsg.io/}{EPSG code}. One of:
\itemize{
\item \code{"4326"}: \href{https://epsg.io/4326}{WGS84}.
\item \code{"3035"}: \href{https://epsg.io/3035}{ETRS89 / ETRS-LAEA}.
\item \code{"3857"}: \href{https://epsg.io/3857}{Pseudo-Mercator}.
}}

\item{cache}{A logical value indicating whether to cache results. Defaults
to \code{TRUE}. See the \strong{Caching strategies} section in
\code{\link[=gisco_set_cache_dir]{gisco_set_cache_dir()}}.}

\item{update_cache}{A logical value indicating whether to refresh the
cached file. Defaults to \code{FALSE}. When set to \code{TRUE}, it forces a new
download.}

\item{cache_dir}{A character string with a path to a cache directory. See
the \strong{Caching strategies} section in \code{\link[=gisco_set_cache_dir]{gisco_set_cache_dir()}}.}

\item{verbose}{A logical value indicating whether to display informational
messages.}

\item{resolution}{A character string or numeric value with the geospatial
data resolution. One of:
\itemize{
\item \code{"60"}: 1:60 million.
\item \code{"20"}: 1:20 million.
\item \code{"10"}: 1:10 million.
\item \code{"03"}: 1:3 million.
\item \code{"01"}: 1:1 million.
}}

\item{spatialtype}{A character string with the type of geometry to return.
Available values are:
\itemize{
\item \code{"RG"}: Regions - \code{MULTIPOLYGON/POLYGON} object.
\item \code{"LB"}: Labels - \code{POINT} object.
\item \code{"BN"}: Boundaries - \code{LINESTRING} object.

Arguments \code{country}, \code{nuts_level} and \code{nuts_id} are only applied when
\code{spatialtype} is \code{"RG"} or \code{"LB"}.
}}

\item{country}{A character vector of country codes. It can be either a
vector of country names, a vector of ISO 3166-1 alpha-3 country codes or a
vector of Eurostat country codes. See also \code{\link[countrycode:countrycode]{countrycode::countrycode()}}.}

\item{nuts_id}{An optional character vector of NUTS IDs.}

\item{nuts_level}{A character string with the NUTS level. One of \code{0},
\code{1}, \code{2}, \code{3} or \code{all} for all levels.}

\item{ext}{A character string with the extension of the file (default
\code{"gpkg"}). One of
\Sexpr[stage=render,results=rd]{giscoR:::db_values("nuts",
"ext",TRUE)}.}
}
\value{
An \code{\link[sf:st_sf]{sf}} object.
}
\description{
The GISCO statistical unit dataset represents the Nomenclature of
Territorial Units for Statistics (NUTS) and statistical regions using
multipart polygon, polyline and point topology. The NUTS geographical
information is completed by attribute tables and a set of cartographic
help lines to better visualize multipart polygonal regions.

NUTS is a hierarchical system divided into three levels:
\itemize{
\item NUTS 1: major socio-economic regions.
\item NUTS 2: basic regions for the application of regional policies.
\item NUTS 3: small regions for specific diagnoses.
}

There is also a NUTS 0 level, which usually corresponds to national
boundaries.

Downloads data from the aggregated GISCO NUTS file, which contains data for
all countries at the requested NUTS level or levels. To download single-unit
NUTS files, use \code{\link[=gisco_get_unit_nuts]{gisco_get_unit_nuts()}}.
}
\details{
The NUTS nomenclature is a hierarchical classification of statistical
regions and subdivides the EU economic territory into regions of three
different levels (NUTS 1, 2 and 3, moving respectively from larger to
smaller territorial units). NUTS 1 is the most aggregated level. Additional
country-level NUTS 0 data are also available for countries where the
statistical national level does not coincide with the administrative
boundaries.

NUTS legislation is amended periodically, so GISCO provides multiple
classification vintages. An unofficial NUTS-like classification is defined
for EFTA countries, candidate countries and potential candidates based on
bilateral agreements between Eurostat and the respective statistical
agencies.

An introduction to the NUTS classification is available here:
\url{https://ec.europa.eu/eurostat/web/nuts/overview}.
}
\section{Copyright}{
See the GISCO statistical unit copyright provisions:
\url{https://ec.europa.eu/eurostat/web/gisco/geodata/statistical-units}.
}

\section{Note}{
Check the download and usage provisions in \code{\link[=gisco_attributions]{gisco_attributions()}}.
}

\examples{
\dontshow{if (requireNamespace("ggplot2", quietly = TRUE)) withAutoprint(\{ # examplesIf}
nuts2 <- gisco_get_nuts(nuts_level = 2)

library(ggplot2)

ggplot(nuts2) +
  geom_sf() +
  # ETRS89 / ETRS-LAEA
  coord_sf(
    crs = 3035, xlim = c(2377294, 7453440),
    ylim = c(1313597, 5628510)
  ) +
  labs(title = "NUTS 2 levels")
# NUTS 3 for Germany.
germany_nuts3 <- gisco_get_nuts(nuts_level = 3, country = "Germany")

ggplot(germany_nuts3) +
  geom_sf() +
  labs(
    title = "NUTS 3 levels",
    subtitle = "Germany",
    caption = gisco_attributions()
  )

# Select specific regions
select_nuts <- gisco_get_nuts(nuts_id = c("ES2", "FRJ", "FRL", "ITC"))

ggplot(select_nuts) +
  geom_sf(aes(fill = CNTR_CODE)) +
  scale_fill_viridis_d()
\dontshow{\}) # examplesIf}
}
\seealso{
\link{gisco_nuts_2024}, \code{\link[eurostat:get_eurostat_geospatial]{eurostat::get_eurostat_geospatial()}}.

See \code{\link[=gisco_bulk_download]{gisco_bulk_download()}} to perform a bulk download of datasets.

See \code{\link[=gisco_get_unit_nuts]{gisco_get_unit_nuts()}} to download single-unit files.

See \code{\link[=gisco_id_api_nuts]{gisco_id_api_nuts()}} to download via the GISCO ID service API.

Statistical unit datasets:
\code{\link[=gisco_get_census]{gisco_get_census()}},
\code{\link[=gisco_get_coastal_lines]{gisco_get_coastal_lines()}},
\code{\link[=gisco_get_lau]{gisco_get_lau()}},
\code{\link[=gisco_get_urban_audit]{gisco_get_urban_audit()}}
}
\concept{stats}
