% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{M49_regions}
\alias{M49_regions}
\title{Dataframe including United Nations M49 geographic regions}
\format{
A data frame object with 250 rows and 9 variables:
\describe{
  \item{ISO3_CODE}{\href{https://en.wikipedia.org/wiki/ISO_3166-1_alpha-3}{ISO 3166-1 alpha-3 code} of each country, as provided by GISCO}
  \item{NAME}{Country name}
  \item{REGION_CODE}{Numeric region code}
  \item{REGION}{Region}
  \item{SUBREGION_CODE}{Numeric sub-region code}
  \item{SUBREGION}{Sub-Region}
  \item{INTERREGION_CODE}{Numeric intermediate Region code}
  \item{INTERREGION}{Intermediate Region}
  \item{DEVELOPED}{Indicates wheter a country is considered "Developed" of "Developing"}
}
}
\source{
\url{https://unstats.un.org/unsd/methodology/m49/#geo-regions}
}
\description{
A dataframe containing geographic regions information, as provided by the UN Standard Country or Area Codes for Statistical Use (M49).
}
\note{
This data was extracted on 29Jan2020 using the reop \url{https://github.com/dieghernan/Country-Codes-and-International-Organizations/}
}
