% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tgs00026}
\alias{tgs00026}
\title{Disposable income of private households by NUTS 2 regions}
\format{
data_frame
}
\source{
\url{https://ec.europa.eu/eurostat/web/products-datasets?p_l_id=4566731&p_v_l_s_g_id=0}
}
\description{
Auxiliary Data Sets
}
\note{
Ported from \code{eurostat} package.
}
