% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gisco_bulk_download.R
\name{gisco_bulk_download}
\alias{gisco_bulk_download}
\title{Bulk download from GISCO API}
\source{
\url{https://gisco-services.ec.europa.eu/distribution/v2/}
}
\usage{
gisco_bulk_download(
  id_giscor = c("countries", "coastal_lines", "communes", "lau", "nuts", "urban_audit",
    "postal_codes"),
  year = 2016,
  cache_dir = NULL,
  update_cache = FALSE,
  verbose = FALSE,
  resolution = 10,
  ext = c("shp", "geojson"),
  recursive = deprecated(),
  ...
)
}
\arguments{
\item{id_giscor}{Type of dataset to be downloaded. Values supported are:
\itemize{
\item \code{"coastallines"}.
\item \code{"communes"}.
\item \code{"countries"}.
\item \code{"lau"}.
\item \code{"nuts"}.
\item \code{"urban_audit"}.
}}

\item{year}{Release year of the file. See \strong{Details}.}

\item{cache_dir}{character string. A path to a cache directory. See
\strong{Caching strategies} section in \code{\link[=gisco_set_cache_dir]{gisco_set_cache_dir()}}.}

\item{update_cache}{logical. Should the cached file be refreshed?. Default
is \code{FALSE}. When set to \code{TRUE} it would force a new download.}

\item{verbose}{logical. If \code{TRUE} displays informational messages.}

\item{resolution}{character string or number. Resolution of the geospatial
data. One of:
\itemize{
\item \code{"60"}: 1:60million
\item \code{"20"}: 1:20million
\item \code{"10"}: 1:10million
\item \code{"03"}: 1:3million
\item \code{"01"}: 1:1million
}}

\item{ext}{Extension of the file(s) to be downloaded. Formats available are
\code{"geojson"}, \code{"shp"}, \code{"svg"}, \code{"json"}, \code{"gdb"}. See \strong{Details}.}

\item{recursive}{Tries to unzip recursively the zip files (if any) included
in the initial bulk download (case of \code{ext = "shp"}).}

\item{...}{Ignored.}
}
\value{
Silent function.
}
\description{
Downloads zipped data from GISCO and extract them on the
\code{\link[=gisco_set_cache_dir]{cache_dir}} folder.
}
\details{
See the years available in the corresponding functions:
\itemize{
\item \code{\link[=gisco_get_coastal_lines]{gisco_get_coastal_lines()}}.
\item \code{\link[=gisco_get_communes]{gisco_get_communes()}}.
\item \code{\link[=gisco_get_countries]{gisco_get_countries()}}.
\item \code{\link[=gisco_get_lau]{gisco_get_lau()}}.
\item \code{\link[=gisco_get_nuts]{gisco_get_nuts()}}.
\item \code{\link[=gisco_get_urban_audit]{gisco_get_urban_audit()}}.
}

The usual extension used across \CRANpkg{giscoR} is \code{"geojson"},
however other formats are already available on GISCO.
}
\examples{
\dontshow{if (gisco_check_access()) withAutoprint(\{ # examplesIf}
\dontrun{

# Write on temp dir
tmp <- file.path(tempdir(), "testexample")

ss <- gisco_bulk_download(
  id_giscor = "countries", resolution = "60",
  year = 2016, ext = "geojson",
  cache_dir = tmp
)
# Read one
library(sf)
f <- list.files(tmp, recursive = TRUE, full.names = TRUE)
f[1]
sf::read_sf(f[1])

# Clean
unlink(tmp, force = TRUE)
}
\dontshow{\}) # examplesIf}
}
\seealso{
Other political: 
\code{\link{gisco_get_units}()}
}
\concept{political}
