% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gisco_check_access.R
\name{gisco_check_access}
\alias{gisco_check_access}
\title{Check access to GISCO API}
\usage{
gisco_check_access()
}
\value{
a logical.
}
\description{
Check if \strong{R} has access to resources at
\url{https://gisco-services.ec.europa.eu/distribution/v2/}.
}
\examples{

gisco_check_access()
}
\seealso{
Other helper: 
\code{\link{gisco_attributions}()}
}
\concept{helper}
