% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gisco_data.R
\docType{data}
\name{gisco_coastallines}
\alias{gisco_coastallines}
\title{World coastal lines \code{POLYGON} object}
\format{
A \code{POLYGON} data frame (resolution: 1:20million, EPSG:4326) object with
3 variables:
\itemize{
\item \strong{FID}
\item \strong{COAS_ID}
\item \strong{geometry}: geometry field
}
}
\source{
\href{https://gisco-services.ec.europa.eu/distribution/v2/coas/geojson/}{\code{COAS_RG_20M_2016_4326.geojson}} file.
}
\description{
A \code{sf} object as provided by GISCO (2016 version).
}
\examples{

coasts <- gisco_coastallines

library(tmap)

tmap_style("classic")

qtm(coasts) + tm_compass(
  position = c("left", "bottom"),
  color.dark = "chocolate"
)
}
\seealso{
\code{\link[=gisco_get_coastallines]{gisco_get_coastallines()}}
}
\concept{dataset}
