% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gisco_coastallines}
\alias{gisco_coastallines}
\title{World coastal lines \code{POLYGON} object}
\format{
A \code{POLYGON} \code{\link[sf:sf]{sf}} object (resolution: 1:20million, EPSG:4326)
with 3 variables:
\describe{
\item{COAS_ID}{Coast ID.}
\item{FID}{FID.}
\item{geometry}{geometry field.}
}
}
\source{
\href{https://gisco-services.ec.europa.eu/distribution/v2/coas/geojson/}{COAS_RG_20M_2016_4326.geojson} file.
}
\description{
A \code{\link[sf:sf]{sf}} object as provided by GISCO (2016 version).
}
\examples{
data("gisco_coastallines")
head(gisco_coastallines)
}
\seealso{
\code{\link[=gisco_get_coastallines]{gisco_get_coastallines()}}

Other dataset: 
\code{\link{gisco_countries}},
\code{\link{gisco_countrycode}},
\code{\link{gisco_db}},
\code{\link{gisco_nuts}}
}
\concept{dataset}
