% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{gisco_countries}
\alias{gisco_countries}
\title{World countries \code{POLYGON} object}
\format{
A \code{MULTIPOLYGON} data frame (resolution: 1:20million, EPSG:4326) object
with 257 rows and 7 variables:
\itemize{
\item \strong{id}: row ID
\item \strong{CNTR_NAME}: Official country name on local language
\item \strong{ISO3_CODE}: ISO 3166-1 alpha-3 code of each country, as provided by
GISCO
\item \strong{CNTR_ID}: Country ID
\item \strong{NAME_ENGL}: Country name in English
\item \strong{FID}: FID
\item \strong{geometry}: geometry field
}
}
\source{
\href{https://gisco-services.ec.europa.eu/distribution/v2/countries/geojson/}{\code{CNTR_RG_20M_2016_4326.geojson}} file.
}
\description{
A \code{sf} object including all
countries as provided by GISCO (2016 version).
}
\examples{

cntry <- gisco_countries
GBR <- subset(cntry, ISO3_CODE == "GBR")

library(ggplot2)

ggplot(GBR) +
  geom_sf(color = "red3", fill = "blue4") +
  theme_void()
}
\seealso{
\code{\link[=gisco_get_countries]{gisco_get_countries()}}
}
\concept{dataset}
