% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{gisco_countries_2024}
\alias{gisco_countries_2024}
\title{World countries \code{POLYGON} \code{\link[sf:sf]{sf}} object 2024}
\format{
A \code{MULTIPOLYGON} data frame (resolution: 1:20million, EPSG:4326) object
with 260 rows and 7 variables:
\describe{
\item{CNTR_ID}{Country ID as per Eurostat.}
\item{CNTR_NAME}{Official country name on local language.}
\item{NAME_ENGL}{Country name in English.}
\item{NAME_FREN}{Country name in French.}
\item{ISO3_CODE}{ISO 3166-1 alpha-3 code of each country, as provided by
GISCO.}
\item{SVRG_UN}{Sovereign status as per United Nations.}
\item{CAPT}{Capitol city.}
\item{EU_STAT}{European Union member.}
\item{EFTA_STAT}{EFTA member.}
\item{CC_STAT}{EU candidate member.}
\item{NAME_GERM}{Country name in German.}
\item{geometry}{geometry field.}
}
}
\source{
\href{https://gisco-services.ec.europa.eu/distribution/v2/countries/gpkg/}{CNTR_RG_20M_2024_4326.gpkg} file.
}
\description{
A \code{\link[sf:sf]{sf}} object including all countries as provided by
GISCO (2024 version).
}
\examples{

data("gisco_countries_2024")
head(gisco_countries_2024)

}
\seealso{
\code{\link[=gisco_get_countries]{gisco_get_countries()}}

Other dataset: 
\code{\link{gisco_coastallines}},
\code{\link{gisco_countries}},
\code{\link{gisco_countrycode}},
\code{\link{gisco_db}},
\code{\link{gisco_nuts}}
}
\concept{dataset}
