% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gisco_data.R
\docType{data}
\name{gisco_db}
\alias{gisco_db}
\title{GISCO database}
\format{
A data frame
}
\source{
GISCO API \code{datasets.json}.
}
\description{
Database with the list of files that the package can load.
}
\details{
This dataframe is used to check the validity of the API calls.
}
\examples{
data(gisco_db)
}
