% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gisco_db}
\alias{gisco_db}
\title{GISCO database}
\format{
A data frame
}
\source{
GISCO API \code{datasets.json}.
}
\description{
Database with the list of files that the package can load.
}
\details{
This data frame is used to check the validity of the API calls.
}
\examples{

data("gisco_db")
}
\seealso{
Other dataset: 
\code{\link{gisco_coastallines}},
\code{\link{gisco_countries}},
\code{\link{gisco_countrycode}},
\code{\link{gisco_nuts}}
}
\concept{dataset}
\keyword{internal}
