% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gisco-get-airports.R
\encoding{UTF-8}
\name{gisco_get_airports}
\alias{gisco_get_airports}
\title{Airports dataset}
\source{
\url{https://ec.europa.eu/eurostat/web/gisco/geodata/transport-networks}.
}
\usage{
gisco_get_airports(
  year = c(2024, 2013, 2006),
  country = NULL,
  cache_dir = NULL,
  update_cache = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{year}{A character string or numeric value with the release year of the
file. One of \code{2024}, \code{2013}, \code{2006}.}

\item{country}{A character vector of country codes. It can be either a
vector of country names, a vector of ISO 3166-1 alpha-3 country codes or a
vector of Eurostat country codes. See also \code{\link[countrycode:countrycode]{countrycode::countrycode()}}.}

\item{cache_dir}{A character string with a path to a cache directory. See
\strong{Caching strategies} section in \code{\link[=gisco_set_cache_dir]{gisco_set_cache_dir()}}.}

\item{update_cache}{A logical value indicating whether to refresh the
cached file. Defaults to \code{FALSE}. When set to \code{TRUE}, it forces a new
download.}

\item{verbose}{A logical value indicating whether to display informational
messages.}
}
\value{
A \code{\link[sf:st_sf]{sf}} object.
}
\description{
This function accesses the GISCO airport and heliport datasets. Airports are
identified using International Civil Aviation Organization (ICAO) airport
codes.
}
\details{
The returned object is transformed to \href{https://epsg.io/4326}{EPSG:4326}.
}
\section{Copyright}{
See the Eurostat general copyright and license provisions:
\url{https://ec.europa.eu/eurostat/web/gisco/geodata}.
}

\examples{
\dontshow{if (gisco_check_access()) withAutoprint(\{ # examplesIf}
airp <- gisco_get_airports(year = 2024)
coast <- giscoR::gisco_get_countries(year = 2024)

if (!is.null(airp) && requireNamespace("ggplot2", quietly = TRUE)) {
  library(ggplot2)

  ggplot(coast) +
    geom_sf(fill = "grey10", color = "grey20") +
    geom_sf(
      data = airp, color = "#00F0FF",
      size = 0.2, alpha = 0.25
    ) +
    theme_void() +
    theme(
      plot.background = element_rect(fill = "black"),
      text = element_text(color = "white"),
      panel.grid = element_blank(),
      plot.title = element_text(face = "bold", hjust = 0.5),
      plot.subtitle = element_text(face = "italic", hjust = 0.5)
    ) +
    labs(
      title = "Airports in Europe", subtitle = "Year 2024",
      caption = "Source: Eurostat, Airports 2024 dataset."
    ) +
    # Center on Europe with EPSG 3035.
    coord_sf(
      crs = 3035,
      xlim = c(2377294, 7453440),
      ylim = c(1313597, 5628510)
    )
}
\dontshow{\}) # examplesIf}
}
\seealso{
Transport network datasets:
\code{\link[=gisco_get_ports]{gisco_get_ports()}}
}
\concept{transport}
