% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gisco-get-airports.R
\encoding{UTF-8}
\name{gisco_get_airports}
\alias{gisco_get_airports}
\title{Airports dataset}
\source{
\url{https://ec.europa.eu/eurostat/web/gisco/geodata/transport-networks}

Copyright:
\url{https://ec.europa.eu/eurostat/web/gisco/geodata}
}
\usage{
gisco_get_airports(
  year = c(2013, 2006),
  country = NULL,
  cache_dir = NULL,
  update_cache = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{year}{character string or number. Release year of the file. One of
\code{2013}, \code{2006}.}

\item{country}{character vector of country codes. It could be either a
vector of country names, a vector of ISO3 country codes or a vector of
Eurostat country codes. See also \code{\link[countrycode:countrycode]{countrycode::countrycode()}}.}

\item{cache_dir}{character string. A path to a cache directory. See
\strong{Caching strategies} section in \code{\link[=gisco_set_cache_dir]{gisco_set_cache_dir()}}.}

\item{update_cache}{logical. Should the cached file be refreshed?. Default
is \code{FALSE}. When set to \code{TRUE} it would force a new download.}

\item{verbose}{logical. If \code{TRUE} displays informational messages.}
}
\value{
A \code{\link[sf:sf]{sf}} object.
}
\description{
This dataset includes the location of over 11,800 Pan European airports and
heliports. The airports are identified using the International Civil
Aviation Organisation (ICAO) airport codes.
}
\details{
Dataset includes objects in \href{https://epsg.io/4326}{EPSG:4326}.
}
\examples{
\dontshow{if (gisco_check_access()) withAutoprint(\{ # examplesIf}
library(sf)

greece <- gisco_get_countries(country = "EL", resolution = 3)
airp_gc <- gisco_get_airports(2013, country = "EL")

library(ggplot2)

if (inherits(airp_gc, "sf")) {
  ggplot(greece) +
    geom_sf(fill = "grey80") +
    geom_sf(data = airp_gc, color = "blue") +
    labs(
      title = "Airports on Greece",
      shape = NULL,
      color = NULL,
      caption = gisco_attributions()
    )
}
\dontshow{\}) # examplesIf}
}
\seealso{
Other transport networks datasets:
\code{\link{gisco_get_ports}()}
}
\concept{transport}
