% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gisco-get-census.R
\name{gisco_get_census}
\alias{gisco_get_census}
\title{Census data}
\source{
\url{https://ec.europa.eu/eurostat/web/gisco/geodata/statistical-units/census}
}
\usage{
gisco_get_census(
  year = "2011",
  cache_dir = NULL,
  update_cache = FALSE,
  verbose = FALSE,
  spatialtype = c("RG", "PT")
)
}
\arguments{
\item{year}{Year of reference. Currently only \code{"2011"} is supported}

\item{cache_dir}{A path to a cache directory. See \strong{About caching}.}

\item{update_cache}{A logical whether to update cache. Default is \code{FALSE}.
When set to \code{TRUE} it would force a fresh download of the source
\code{.gpkg} file.}

\item{verbose}{Logical, displays information. Useful for debugging,
default is \code{FALSE}.}

\item{spatialtype}{Type of geometry to be returned:
\itemize{
\item \code{"PT"}: Points - \code{POINT} object.
\item \code{"RG"}: Regions - \code{MULTIPOLYGON/POLYGON} object.
}}
}
\value{
A \code{\link[sf:sf]{sf}} object specified by \code{spatialtype}.
}
\description{
This data set shows pan European communal boundaries depicting the situation
at the corresponding Census.
}
\section{About caching}{
You can set your \code{cache_dir} with \code{\link[=gisco_set_cache_dir]{gisco_set_cache_dir()}}.

Sometimes cached files may be corrupt. On that case, try re-downloading
the data setting \code{update_cache = TRUE}.

If you experience any problem on download, try to download the
corresponding file by any other method and save it on your
\code{cache_dir}. Use the option \code{verbose = TRUE} for debugging the API query.

For a complete list of files available check \link{gisco_db}.
}

\examples{
\dontshow{if (gisco_check_access()) withAutoprint(\{ # examplesIf}
\donttest{
library(sf)

pts <- gisco_get_census(spatialtype = "PT")

pts
}
\dontshow{\}) # examplesIf}
}
\concept{stats}
