% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gisco-get-coastal-lines.R
\encoding{UTF-8}
\name{gisco_get_coastal_lines}
\alias{gisco_get_coastal_lines}
\alias{gisco_get_coastallines}
\title{Coastal lines dataset}
\source{
\url{https://gisco-services.ec.europa.eu/distribution/v2/}.

Copyright:
\url{https://ec.europa.eu/eurostat/web/gisco/geodata/statistical-units}.
}
\usage{
gisco_get_coastal_lines(
  year = 2016,
  epsg = 4326,
  cache = TRUE,
  update_cache = FALSE,
  cache_dir = NULL,
  verbose = FALSE,
  resolution = 20,
  ext = "gpkg"
)
}
\arguments{
\item{year}{A character string or numeric value with the release year of the
file.
One of \Sexpr[stage=render,results=rd]{giscoR:::db_values("coastal_lines",
"year",TRUE)}.}

\item{epsg}{A character string or numeric value with the coordinate
reference system as a 4-digit \href{https://epsg.io/}{EPSG code}. One of:
\itemize{
\item \code{"4326"}: \href{https://epsg.io/4326}{WGS84}.
\item \code{"3035"}: \href{https://epsg.io/3035}{ETRS89 / ETRS-LAEA}.
\item \code{"3857"}: \href{https://epsg.io/3857}{Pseudo-Mercator}.
}}

\item{cache}{A logical value indicating whether to cache results. Defaults
to \code{TRUE}. See \strong{Caching strategies} section in \code{\link[=gisco_set_cache_dir]{gisco_set_cache_dir()}}.}

\item{update_cache}{A logical value indicating whether to refresh the
cached file. Defaults to \code{FALSE}. When set to \code{TRUE}, it forces a new
download.}

\item{cache_dir}{A character string with a path to a cache directory. See
\strong{Caching strategies} section in \code{\link[=gisco_set_cache_dir]{gisco_set_cache_dir()}}.}

\item{verbose}{A logical value. If \code{TRUE} displays informational messages.}

\item{resolution}{A character string or numeric value with the geospatial
data resolution. One of:
\itemize{
\item \code{"60"}: 1:60 million.
\item \code{"20"}: 1:20 million.
\item \code{"10"}: 1:10 million.
\item \code{"03"}: 1:3 million.
\item \code{"01"}: 1:1 million.
}}

\item{ext}{A character value with the extension of the file (default
\code{"gpkg"}). One of
\Sexpr[stage=render,results=rd]{giscoR:::db_values("coastal_lines",
"ext",TRUE)}.}
}
\value{
A \code{\link[sf:st_sf]{sf}} object.
}
\description{
Download global coastal lines.
}
\section{Note}{
Check the download and usage provisions in \code{\link[=gisco_attributions]{gisco_attributions()}}.
}

\examples{
\dontshow{if (gisco_check_access()) withAutoprint(\{ # examplesIf}
coast <- gisco_get_coastal_lines()

library(ggplot2)

ggplot(coast) +
  geom_sf(color = "#1278AB", fill = "#FDFBEA") +
  # Zoom on the Mediterranean Sea.
  coord_sf(
    xlim = c(-4, 35),
    ylim = c(31, 45)
  ) +
  theme_minimal() +
  theme(
    panel.background = element_rect(fill = "#C7E7FB", color = NA),
    panel.border = element_rect(colour = "black", fill = NA)
  )
\dontshow{\}) # examplesIf}
}
\seealso{
\link{gisco_coastal_lines}.

See \code{\link[=gisco_bulk_download]{gisco_bulk_download()}} to perform a bulk download of datasets.

Statistical unit datasets:
\code{\link[=gisco_get_census]{gisco_get_census()}},
\code{\link[=gisco_get_lau]{gisco_get_lau()}},
\code{\link[=gisco_get_nuts]{gisco_get_nuts()}},
\code{\link[=gisco_get_urban_audit]{gisco_get_urban_audit()}}
}
\concept{stats}
