% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gisco_get_coastallines.R
\name{gisco_get_coastallines}
\alias{gisco_get_coastallines}
\title{Get GISCO coastlines \code{\link[sf:sf]{sf}} polygons}
\source{
\url{https://gisco-services.ec.europa.eu/distribution/v2/}
}
\usage{
gisco_get_coastallines(
  year = "2016",
  epsg = "4326",
  cache = TRUE,
  update_cache = FALSE,
  cache_dir = NULL,
  verbose = FALSE,
  resolution = "20"
)
}
\arguments{
\item{year}{Release year. One of \code{"2006"}, \code{"2010"}, \code{"2013"} or \code{"2016"}.}

\item{epsg}{character string or number. Projection of the map: 4-digit
\href{https://epsg.io/}{EPSG code}. One of:
\itemize{
\item \code{"4326"}: \href{https://epsg.io/4326}{WGS84}
\item \code{"3035"}: \href{https://epsg.io/3035}{ETRS89 / ETRS-LAEA}
\item \code{"3857"}: \href{https://epsg.io/3857}{Pseudo-Mercator}
}}

\item{cache}{logical. Whether to do caching. Default is \code{TRUE}. See
\strong{Caching strategies} section in \code{\link[=gisco_set_cache_dir]{gisco_set_cache_dir()}}.}

\item{update_cache}{.logical. Should the cached file be refreshed?. Default
is \code{FALSE}. When set to \code{TRUE} it would force a new download.}

\item{cache_dir}{character string. A path to a cache directory. See
\strong{Caching strategies} section in \code{\link[=gisco_set_cache_dir]{gisco_set_cache_dir()}}.}

\item{verbose}{logical. If \code{TRUE} displays informational messages.}

\item{resolution}{Resolution of the geospatial data. One of:
\itemize{
\item \code{"60"}: 1:60million
\item \code{"20"}: 1:20million
\item \code{"10"}: 1:10million
\item \code{"03"}: 1:3million
\item \code{"01"}: 1:1million
}}
}
\value{
A \code{\link[sf:sf]{sf}} \code{POLYGON} object.
}
\description{
Downloads worldwide coastlines
}
\note{
Please check the download and usage provisions on \code{\link[=gisco_attributions]{gisco_attributions()}}.
}
\examples{
\dontshow{if (gisco_check_access()) withAutoprint(\{ # examplesIf}
coast <- gisco_get_coastallines()

library(ggplot2)

ggplot(coast) +
  geom_sf(color = "#1278AB", fill = "#FDFBEA") +
  # Zoom on Caribe
  coord_sf(
    xlim = c(-99, -49),
    ylim = c(4, 30)
  ) +
  theme_minimal() +
  theme(
    panel.background = element_rect(fill = "#C7E7FB", color = NA),
    panel.border = element_rect(colour = "black", fill = NA)
  )
\dontshow{\}) # examplesIf}
}
\seealso{
\link{gisco_coastallines}

Other statistical units datasets:
\code{\link{gisco_get_census}()},
\code{\link{gisco_get_nuts}()}
}
\concept{stats}
