% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gisco_get_communes.R
\name{gisco_get_communes}
\alias{gisco_get_communes}
\title{Download Geospatial Communes Data from GISCO}
\source{
\href{https://gisco-services.ec.europa.eu/distribution/v2/communes/}{GISCO Communes}
}
\usage{
gisco_get_communes(
  year = "2016",
  epsg = "4326",
  update_cache = FALSE,
  cache_dir = NULL,
  spatialtype = "RG",
  country_iso3 = NULL
)
}
\arguments{
\item{year}{Release year. One of
"2001", "2004", "2006", "2008", "2010", "2013" or 2016}

\item{epsg}{projection of the map: 4-digit \href{https://spatialreference.org/ref/epsg/}{EPSG code}. One of:
\itemize{
\item "4326" - WGS84
\item "3035" - ETRS89 / ETRS-LAEA
\item "3857" - Pseudo-Mercator
}}

\item{update_cache}{a logical whether to update cache.}

\item{cache_dir}{a path to a cache directory. The directory have to exist.  The \code{NULL} (default) uses and creates \code{/gisco} directory in the temporary directory from \code{\link{tempdir}}. The directory can also be set with \code{options(gisco_cache_dir = <path>)}.}

\item{spatialtype}{Type of geometry to be returned:
\itemize{
 \item RG: Regions - Multipolygon
 \item LB: Labels - Point
 \item BN: Boundaries - Multilines
 \item COASTL: coastlines - Multilines
 \item INLAND: inland boundaries - Multilines
}}

\item{country_iso3}{Optional. A character vector of ISO-3 country codes. See Details.}
}
\value{
a \code{sf} object.
}
\description{
Downloads a simple feature (\code{sf}) object.
}
\details{
\code{country_iso3} only available on specific datasets. Some \code{spatialtype} datasets (as Multilines data-types) may not have country-level identifies.

You can convert Eurostat country codes to ISO3 codes using the \code{\link[countrycode]{countrycode}} function:

\code{eurostat_codes <- c("ES","UK","EL","PL","PT")}

\code{countrycode::countrycode(eurostat_codes,origin = "eurostat", destination = "iso3c")}


If you experience any problem on download, try to download the file by any other method and set \code{cache_dir = <folder>}.
}
\note{
Please check the download and usage provisions on \link{gisco_attributions}.
}
\examples{
\donttest{
library(sf)

benelux <- c("BEL", "NLD", "LUX")
communes <- gisco_get_communes(country_iso3 = benelux)

plot(
  communes[, "CNTR_ID"],
  pal = c("black", "#00A3E0", "orange"),
  border = "grey90",
  main = "Communes on Benelux (2016)",
  key.pos = NULL
)
title(sub = gisco_attributions())
}
}
\author{
dieghernan, \url{https://github.com/dieghernan/}
}
