% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gisco_get_countries.R
\name{gisco_get_countries}
\alias{gisco_get_countries}
\title{Download Geospatial Country Data from GISCO}
\source{
\href{https://gisco-services.ec.europa.eu/distribution/v2/countries/}{GISCO Countries}
}
\usage{
gisco_get_countries(
  resolution = "20",
  year = "2016",
  crs = "4326",
  cache = TRUE,
  update_cache = FALSE,
  cache_dir = NULL,
  spatialtype = "RG",
  country = NULL,
  region = NULL
)
}
\arguments{
\item{resolution}{Resolution of the geospatial data. One of
\itemize{
   \item "60" (1:60million),
   \item "20" (1:20million)
   \item "10" (1:10million)
   \item "03" (1:3million) or
   \item "01" (1:1million).
   }}

\item{year}{Release year. One of
"2001", "2006", "2010", "2013", "2016" or "2020"}

\item{crs}{projection of the map: 4-digit \href{http://spatialreference.org/ref/epsg/}{EPSG code}. One of:
\itemize{
\item "4326" - WGS84
\item "3035" - ETRS89 / ETRS-LAEA
\item "3857" - Pseudo-Mercator
}}

\item{cache}{a logical whether to do caching. Default is \code{TRUE}.}

\item{update_cache}{a logical whether to update cache.}

\item{cache_dir}{a path to a cache directory. The directory have to exist.  The \code{NULL} (default) uses and creates \code{/gisco} directory in the temporary directory from \code{\link{tempdir}}. The directory can also be set with \code{options(gisco_cache_dir = <path>)}.}

\item{spatialtype}{Type of geometry to be returned:
\itemize{
 \item RG: Regions - Multipolygon
 \item LB: Labels - Point
 \item BN: Boundaries - Multilines
 \item COASTL: coastlines - Multilines
 \item INLAND: inland boundaries - Multilines
}}

\item{country}{Optional. A character vector of ISO-3 country codes. See Details}

\item{region}{Optional. A character vector of UN M49 region codes. Possible values are "Africa", "Americas", "Asia", "Europe", "Oceania". See Details and \link{gisco_countrycode}}
}
\value{
a \code{sf} object.
}
\description{
Downloads a simple feature (\code{sf}) object.
}
\details{
\code{country} and \code{region} only available when applicable.
Some \code{spatialtype} datasets (as Multilines data-types) may not have country-level identifies.
}
\note{
COPYRIGHT NOTICE

When data downloaded from this page
\url{http://ec.europa.eu/eurostat/web/gisco/geodata/reference-data/administrative-units-statistical-units}
is used in any printed or electronic publication,
in addition to any other provisions
applicable to the whole Eurostat website,
data source will have to be acknowledged
in the legend of the map and
in the introductory page of the publication
with the following copyright notice:
\itemize{
	\item EN: (C) EuroGeographics for the administrative boundaries
	\item FR: (C) EuroGeographics pour les limites administratives
	\item DE: (C) EuroGeographics bezuglich der Verwaltungsgrenzen
}
For publications in languages other than
English, French or German,
the translation of the copyright notice
in the language of the publication shall be used.

If you intend to use the data commercially,
please contact EuroGeographics for
information regarding their licence agreements.
}
\examples{
library(sf)

# Some data are already available for speed up the process
opar <- par(no.readonly = TRUE)
par(mar = c(2, 0, 0, 0))
africa2016 <-  gisco_get_countries(region = "Africa")
angola_namibia <-  gisco_get_countries(country = c("AGO", "NAM"))

plot(st_geometry(africa2016), bg = "#C6ECFF", col = NA)
plot(st_geometry(gisco_countries_20M_2016),
     col = "#E0E0E0",
     add = TRUE)
plot(st_geometry(africa2016), col = "#F6E1B9", add = TRUE)
plot(st_geometry(angola_namibia), col = "#FEFEE9", add = TRUE)

mtext(gisco_attributions(), side = 1, cex = 0.8)

# Change crs and resolution

cntries2020 <-
  gisco_get_countries(year = "2020",
                      crs = "3035",
                      resolution = "60")
plot(st_geometry(cntries2020), bg = "#C6ECFF", col = "#E0E0E0")
mtext(gisco_attributions(), side = 1, cex = 0.8)
par(opar)

# Several geometry types
coastl <-
  gisco_get_countries(spatialtype = "COASTL", resolution = "60")

inland <-
  gisco_get_countries(spatialtype = "INLAND", resolution = "60")
}
\seealso{
\link{gisco_countrycode}
}
\author{
dieghernan, \url{https://github.com/dieghernan/}
}
