% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gisco_get_countries.R
\name{gisco_get_countries}
\alias{gisco_get_countries}
\title{Download Geospatial Country Data from GISCO}
\source{
\href{https://gisco-services.ec.europa.eu/distribution/v2/countries/}{GISCO Countries}
}
\usage{
gisco_get_countries(
  resolution = "60",
  year = "2016",
  epsg = "4326",
  cache = TRUE,
  update_cache = FALSE,
  cache_dir = NULL,
  spatialtype = "RG",
  country_iso3 = NULL,
  region = NULL
)
}
\arguments{
\item{resolution}{Resolution of the geospatial data. One of
\itemize{
   \item "60" (1:60million),
   \item "20" (1:20million)
   \item "10" (1:10million)
   \item "03" (1:3million) or
   \item "01" (1:1million).
   }}

\item{year}{Release year. One of
"2001", "2006", "2010", "2013", "2016" or "2020"}

\item{epsg}{projection of the map: 4-digit \href{http://spatialreference.org/ref/epsg/}{EPSG code}. One of:
\itemize{
\item "4326" - WGS84
\item "3035" - ETRS89 / ETRS-LAEA
\item "3857" - Pseudo-Mercator
}}

\item{cache}{a logical whether to do caching. Default is \code{TRUE}.}

\item{update_cache}{a logical whether to update cache.}

\item{cache_dir}{a path to a cache directory. The directory have to exist.  The \code{NULL} (default) uses and creates \code{/gisco} directory in the temporary directory from \code{\link{tempdir}}. The directory can also be set with \code{options(gisco_cache_dir = <path>)}.}

\item{spatialtype}{Type of geometry to be returned:
\itemize{
 \item RG: Regions - Multipolygon
 \item LB: Labels - Point
 \item BN: Boundaries - Multilines
 \item COASTL: coastlines - Multilines
 \item INLAND: inland boundaries - Multilines
}}

\item{country_iso3}{Optional. A character vector of ISO-3 country codes. See Details}

\item{region}{Optional. A character vector of UN M49 region codes. Possible values are "Africa", "Americas", "Asia", "Europe", "Oceania". See Details and \link{gisco_countrycode}}
}
\value{
a \code{sf} object.
}
\description{
Downloads a simple feature (\code{sf}) object.
}
\details{
\code{country_iso3} and \code{region} only available when applicable.
You can convert Eurostat country codes to ISO3 codes using the \code{\link[countrycode]{countrycode}} function:

\code{eurostat_codes <- c("ES","UK","EL","PL","PT")}

\code{countrycode::countrycode(eurostat_codes,origin = "eurostat", destination = "iso3c")}
}
\note{
Please check the download and usage provisions on \link{gisco_attributions}.
}
\examples{
library(sf)

# Different resolutions
DNK_res60 <- gisco_get_countries(country_iso3 = "DNK")
DNK_res20 <-
  gisco_get_countries(resolution = "20", country_iso3 = "DNK")
DNK_res10 <-
  gisco_get_countries(resolution = "10", country_iso3 = "DNK")
DNK_res03 <-
  gisco_get_countries(resolution = "03", country_iso3 = "DNK")

opar <- par(no.readonly = TRUE)
par(mfrow = c(2, 2), mar = c(3, 0, 2, 0))
plot(st_geometry(DNK_res60), col = "red", main = "60M")
plot(st_geometry(DNK_res20), col = "blue", main = "20M")
plot(st_geometry(DNK_res10), col = "grey50", main = "10M")
plot(st_geometry(DNK_res03), col = "black", main = "03M")
title(sub = gisco_attributions(), line = 1)
par(opar)

# Different projections
cntr_4326 <- gisco_get_countries(epsg = "4326")
cntr_3857 <- gisco_get_countries(epsg = "3857")
cntr_3035 <- gisco_get_countries(epsg = "3035")

opar <- par(no.readonly = TRUE)
par(mfrow = c(1, 3), mar = c(3, 0, 0, 0))
plot(st_geometry(cntr_4326), graticule = TRUE)
plot(
  st_geometry(cntr_3857),
  graticule = TRUE,
  ylim = c(-13000000, 13000000)
)
plot(st_geometry(cntr_3035), graticule = TRUE)
title(sub = gisco_attributions(), line = 1)
par(opar)

# Labels and Lines available
coast <- gisco_get_countries(spatialtype = "COASTL")
labs <- gisco_get_countries(spatialtype = "LB")

plot(st_geometry(coast), col = "blue1")
plot(st_geometry(labs),
     col = "red",
     add = TRUE,
     pch = 19)

title(sub = gisco_attributions(copyright = FALSE), line = 1)
}
\seealso{
\link{gisco_countrycode}
}
\author{
dieghernan, \url{https://github.com/dieghernan/}
}
