% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gisco_get_countries.R
\name{gisco_get_countries}
\alias{gisco_get_countries}
\title{Download Geospatial Country Data from GISCO}
\source{
\href{https://gisco-services.ec.europa.eu/distribution/v2/countries/}{GISCO Countries}
}
\usage{
gisco_get_countries(
  resolution = "60",
  year = "2016",
  epsg = "4326",
  cache = TRUE,
  update_cache = FALSE,
  cache_dir = NULL,
  spatialtype = "RG",
  country_iso3 = NULL,
  region = NULL
)
}
\arguments{
\item{resolution}{Resolution of the geospatial data. One of
\itemize{
   \item "60" (1:60million),
   \item "20" (1:20million)
   \item "10" (1:10million)
   \item "03" (1:3million) or
   \item "01" (1:1million).
   }}

\item{year}{Release year. One of
"2001", "2006", "2010", "2013", "2016" or "2020"}

\item{epsg}{projection of the map: 4-digit \href{https://spatialreference.org/ref/epsg/}{EPSG code}. One of:
\itemize{
\item "4326" - WGS84
\item "3035" - ETRS89 / ETRS-LAEA
\item "3857" - Pseudo-Mercator
}}

\item{cache}{a logical whether to do caching. Default is \code{TRUE}.}

\item{update_cache}{a logical whether to update cache.}

\item{cache_dir}{a path to a cache directory. The directory have to exist.  The \code{NULL} (default) uses and creates \code{/gisco} directory in the temporary directory from \code{\link{tempdir}}. The directory can also be set with \code{options(gisco_cache_dir = <path>)}.}

\item{spatialtype}{Type of geometry to be returned:
\itemize{
 \item RG: Regions - Multipolygon
 \item LB: Labels - Point
 \item BN: Boundaries - Multilines
 \item COASTL: coastlines - Multilines
 \item INLAND: inland boundaries - Multilines
}}

\item{country_iso3}{Optional. A character vector of ISO-3 country codes. See Details}

\item{region}{Optional. A character vector of UN M49 region codes. Possible values are "Africa", "Americas", "Asia", "Europe", "Oceania". See Details and \link{gisco_countrycode}}
}
\value{
a \code{sf} object.
}
\description{
Downloads a simple feature (\code{sf}) object.
}
\details{
\code{country_iso3} and \code{region} only available when applicable.
You can convert Eurostat country codes to ISO3 codes using the \code{\link[countrycode]{countrycode}} function:

eurostat_codes <- c("ES","UK","EL","PL","PT")\cr
\cr
countrycode::countrycode(\cr
  eurostat_codes,\cr
  origin = "eurostat",\cr
  destination = "iso3c"\cr
)
}
\note{
Please check the download and usage provisions on \link{gisco_attributions}.
}
\examples{

library(sf)

sf_world <- gisco_get_countries()
sf_africa <- gisco_get_countries(region = 'Africa')
sf_benelux <-
  gisco_get_countries(resolution = "20",
                      country_iso3 = c('BEL', 'NLD', 'LUX'))

plot(st_geometry(sf_world), col = "seagreen2")
title(sub = gisco_attributions(), line = 1)

plot(st_geometry(sf_africa),
     col = c("springgreen4", "darkgoldenrod1", "red2"))
title(sub = gisco_attributions(), line = 1)

plot(st_geometry(sf_benelux),
     col = c("grey10", "deepskyblue2", "orange"))
title(sub = gisco_attributions(), line = 1)
library(sf)
}
\seealso{
\link{gisco_countrycode}
}
\author{
dieghernan, \url{https://github.com/dieghernan/}
}
