% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gisco_get_grid.R
\name{gisco_get_grid}
\alias{gisco_get_grid}
\title{Get the grid cells covering the European land territory, for various resolutions.}
\source{
\href{https://ec.europa.eu/eurostat/web/gisco/geodata/reference-data/grids}{GISCO API Grids}
}
\usage{
gisco_get_grid(
  resolution = "20",
  spatialtype = "REGION",
  cache_dir = NULL,
  update_cache = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{resolution}{Resolution of the grid cells on kms. Available values are
\code{1, 2, 5, 10, 20, 50, 100}. See Details}

\item{spatialtype}{Select one of \code{REGION,POINT}}

\item{cache_dir, update_cache, verbose}{See \link{gisco_get}}
}
\value{
A \code{POLYGON/POINT} object.
}
\description{
These datasets contain grid cells covering the European land territory, for various resolutions from 1km to 100km. Base statistics such as population figures are provided for these cells.
}
\details{
Files are distributed on EPSG:3035.

  The file sizes range is from 428K (\code{resolution = "100"})
to 1.7G \code{resolution = "1"}. For resolutions 1km and 2km you would
need to confirm the download.
}
\note{
There are specific downloading provisions, please see \url{https://ec.europa.eu/eurostat/web/gisco/geodata/reference-data/grids}
}
\examples{
library(sf)

grid <- gisco_get_grid()
breaks <- unique(quantile(grid$TOT_P_2006, probs = seq(0, 1, 0.01)))

# Remove coastal grids
grid <- grid[grid$LAND_PC > 0.2, ]

plot(
        grid[, "TOT_P_2011"],
        pal = hcl.colors(length(breaks), "ag_Sunset"),
        key.pos = NULL,
        main = "[Grid] Population on Europe (2011)",
        breaks = breaks,
        xlim = c(2636691, 7315863),
        ylim = c(1422800 , 5410788),
        border = NA
)
}
\author{
dieghernan, \url{https://github.com/dieghernan/}
}
