% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gisco-get-healthcare.R
\name{gisco_get_healthcare}
\alias{gisco_get_healthcare}
\title{Get locations of healthcare services in Europe}
\source{
\url{https://ec.europa.eu/eurostat/web/gisco/geodata/basic-services}
}
\usage{
gisco_get_healthcare(
  year = c("2023", "2020"),
  cache = TRUE,
  update_cache = FALSE,
  cache_dir = NULL,
  verbose = FALSE,
  country = NULL
)
}
\arguments{
\item{year}{Release year of the file. One of \code{"2020"}, \code{"2023"} (default).}

\item{cache}{logical. Whether to do caching. Default is \code{TRUE}. See
\strong{Caching strategies} section in \code{\link[=gisco_set_cache_dir]{gisco_set_cache_dir()}}.}

\item{update_cache}{.logical. Should the cached file be refreshed?. Default
is \code{FALSE}. When set to \code{TRUE} it would force a new download.}

\item{cache_dir}{character string. A path to a cache directory. See
\strong{Caching strategies} section in \code{\link[=gisco_set_cache_dir]{gisco_set_cache_dir()}}.}

\item{verbose}{logical. If \code{TRUE} displays informational messages.}

\item{country}{Optional. A character vector of country codes. It could be
either a vector of country names, a vector of ISO3 country codes or a
vector of Eurostat country codes. See also \code{\link[countrycode:countrycode]{countrycode::countrycode()}}.}
}
\value{
A \code{POINT} \code{\link[sf:sf]{sf}} object.
}
\description{
The dataset contains information on main healthcare services considered to
be 'hospitals' by Member States.
}
\details{
Files are distributed on EPSG:4326. Metadata available on
\url{https://gisco-services.ec.europa.eu/pub/healthcare/metadata.pdf}.
}
\examples{
\dontshow{if (gisco_check_access()) withAutoprint(\{ # examplesIf}
\donttest{

health_be <- gisco_get_healthcare(country = "Belgium")

# Plot if downloaded
if (inherits(health_be, "sf")) {
  library(ggplot2)
  ggplot(health_be) +
    geom_sf()
}
}
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=gisco_get_countries]{gisco_get_countries()}}

Other infrastructure: 
\code{\link{gisco_get_airports}()},
\code{\link{gisco_get_education}()}
}
\author{
dieghernan, \url{https://github.com/dieghernan/}
}
\concept{infrastructure}
