% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gisco_get_lau.R
\name{gisco_get_lau}
\alias{gisco_get_lau}
\title{Download Geospatial Local Administrative Units Data from GISCO}
\source{
\href{https://gisco-services.ec.europa.eu/distribution/v2/lau/}{GISCO Local Administrative Units}
}
\usage{
gisco_get_lau(
  year = "2016",
  crs = "4326",
  update_cache = FALSE,
  cache_dir = NULL,
  country = NULL,
  gisco_id = NULL
)
}
\arguments{
\item{year}{Release year. One of
"2016", "2017", "2018" or "2019"}

\item{crs}{projection of the map: 4-digit \href{http://spatialreference.org/ref/epsg/}{EPSG code}. One of:
\itemize{
\item "4326" - WGS84
\item "3035" - ETRS89 / ETRS-LAEA
\item "3857" - Pseudo-Mercator
}}

\item{update_cache}{a logical whether to update cache.}

\item{cache_dir}{a path to a cache directory. The directory have to exist.  The \code{NULL} (default) uses and creates \code{/gisco} directory in the temporary directory from \code{\link{tempdir}}. The directory can also be set with \code{options(gisco_cache_dir = <path>)}.}

\item{country}{Optional. A character vector of ISO-3 country codes. See Details.}

\item{gisco_id}{Optional. A character vector of GISCO_ID LAU values.}
}
\value{
a \code{POLYGON} object on \code{sf} format.
}
\description{
Downloads a simple feature (\code{sf}) object.
}
\details{
See \url{https://ec.europa.eu/eurostat/web/nuts/local-administrative-units} for more detail about LAUs.

If you experience any problem on download, try to download the file by any other method and set \code{cache_dir = <folder>}.
}
\note{
COPYRIGHT NOTICE

When data downloaded from this page
\url{http://ec.europa.eu/eurostat/web/gisco/geodata/reference-data/administrative-units-statistical-units}
is used in any printed or electronic publication,
in addition to any other provisions
applicable to the whole Eurostat website,
data source will have to be acknowledged
in the legend of the map and
in the introductory page of the publication
with the following copyright notice:
\itemize{
	\item EN: (C) EuroGeographics for the administrative boundaries
	\item FR: (C) EuroGeographics pour les limites administratives
	\item DE: (C) EuroGeographics bezuglich der Verwaltungsgrenzen
}
For publications in languages other than
English, French or German,
the translation of the copyright notice
in the language of the publication shall be used.

If you intend to use the data commercially,
please contact EuroGeographics for
information regarding their licence agreements.
}
\examples{
\donttest{
lau <-  gisco_get_lau(year = "2019")
}
}
\author{
dieghernan, \url{https://github.com/dieghernan/}
}
