% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gisco-get-lau.R
\name{gisco_get_lau}
\alias{gisco_get_lau}
\title{Get GISCO urban areas \code{\link[sf:sf]{sf}} polygons, points and lines}
\usage{
gisco_get_lau(
  year = "2024",
  epsg = "4326",
  cache = deprecated(),
  update_cache = FALSE,
  cache_dir = NULL,
  verbose = FALSE,
  country = NULL,
  gisco_id = NULL
)
}
\arguments{
\item{year}{Release year of the file:
\itemize{
\item For \code{gisco_get_communes()} one of
\Sexpr[stage=render,results=rd]{giscoR:::for_docs("communes",
"year",TRUE)}.
\item For \code{gisco_get_lau()} one of
\Sexpr[stage=render,results=rd]{giscoR:::for_docs("lau",
"year",TRUE)}.
}}

\item{epsg}{projection of the map: 4-digit \href{https://epsg.io/}{EPSG code}.
One of:
\itemize{
\item \code{"4326"}: WGS84
\item \code{"3035"}: ETRS89 / ETRS-LAEA
\item \code{"3857"}: Pseudo-Mercator
}}

\item{cache}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}. These functions always
caches the result due to the size. \code{cache_dir} can be set to
\code{\link[base:tempfile]{base::tempdir()}}, so the file would be deleted when the \strong{R} session is
closed.}

\item{update_cache}{logical. Should the cached file be refreshed?. Default
is \code{FALSE}. When set to \code{TRUE} it would force a new download.}

\item{cache_dir}{character string. A path to a cache directory. See
\strong{Caching strategies} section in \code{\link[=gisco_set_cache_dir]{gisco_set_cache_dir()}}.}

\item{verbose}{logical. If \code{TRUE} displays informational messages.}

\item{country}{character vector of country codes. It could be either a
vector of country names, a vector of ISO3 country codes or a vector of
Eurostat country codes. See also \code{\link[countrycode:countrycode]{countrycode::countrycode()}}.}

\item{gisco_id}{Optional. A character vector of GISCO_ID LAU values.}
}
\value{
A \code{\link[sf:sf]{sf}} object specified by \code{spatialtype}. In the case of
\code{\link[=gisco_get_lau]{gisco_get_lau()}}, a \code{POLYGON} object.
}
\description{
\code{\link[=gisco_get_communes]{gisco_get_communes()}} and \code{\link[=gisco_get_lau]{gisco_get_lau()}} download shapes of Local
Urban Areas, that correspond roughly with towns and cities.
}
\note{
Please check the download and usage provisions on \code{\link[=gisco_attributions]{gisco_attributions()}}.
}
\seealso{
Other statistical units datasets:
\code{\link{gisco_get_census}()},
\code{\link{gisco_get_coastallines}()},
\code{\link{gisco_get_nuts}()}
}
\concept{stats}
