% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gisco-get-ports.R
\encoding{UTF-8}
\name{gisco_get_ports}
\alias{gisco_get_ports}
\title{Ports dataset}
\source{
\url{https://ec.europa.eu/eurostat/web/gisco/geodata/transport-networks}.
}
\usage{
gisco_get_ports(
  year = c(2013, 2009),
  country = NULL,
  cache_dir = NULL,
  update_cache = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{year}{A character string or numeric value with the release year of the
file. One of
\code{2013}, \code{2009}.}

\item{country}{A character vector of country codes. It can be either a
vector of country names, a vector of ISO 3166-1 alpha-3 country codes or a
vector of Eurostat country codes. See also \code{\link[countrycode:countrycode]{countrycode::countrycode()}}.}

\item{cache_dir}{A character string with a path to a cache directory. See
\strong{Caching strategies} section in \code{\link[=gisco_set_cache_dir]{gisco_set_cache_dir()}}.}

\item{update_cache}{A logical value indicating whether to refresh the
cached file. Defaults to \code{FALSE}. When set to \code{TRUE}, it forces a new
download.}

\item{verbose}{A logical value indicating whether to display informational
messages.}
}
\value{
A \code{\link[sf:st_sf]{sf}} object.
}
\description{
This dataset includes port locations worldwide. The ports are identified
following the UN LOCODE list.
}
\details{
The returned object is transformed to \href{https://epsg.io/4326}{EPSG:4326}.

\code{\link[=gisco_get_ports]{gisco_get_ports()}} adds a new field, \code{CNTR_ISO2}, to identify the country
of the port.
}
\section{Copyright}{
See the Eurostat general copyright and license provisions:
\url{https://ec.europa.eu/eurostat/web/gisco/geodata}.
}

\examples{
\dontshow{if (gisco_check_access()) withAutoprint(\{ # examplesIf}
library(sf)

ports <- gisco_get_ports(2013)
coast <- giscoR::gisco_coastal_lines

if (!is.null(ports) && requireNamespace("ggplot2", quietly = TRUE)) {
  library(ggplot2)

  ggplot(coast) +
    geom_sf(fill = "grey10", color = "grey20") +
    geom_sf(
      data = ports, color = "#6bb857",
      size = 0.2, alpha = 0.25
    ) +
    theme_void() +
    theme(
      plot.background = element_rect(fill = "black"),
      text = element_text(color = "white"),
      panel.grid = element_blank(),
      plot.title = element_text(face = "bold", hjust = 0.5),
      plot.subtitle = element_text(face = "italic", hjust = 0.5)
    ) +
    labs(
      title = "Ports worldwide", subtitle = "Year 2013",
      caption = "Source: Eurostat, Ports 2013 dataset."
    ) +
    coord_sf(crs = "ESRI:54030")
}
\dontshow{\}) # examplesIf}
}
\seealso{
Transport network datasets:
\code{\link[=gisco_get_airports]{gisco_get_airports()}}
}
\concept{transport}
