% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gisco-get-urban-audit.R
\encoding{UTF-8}
\name{gisco_get_urban_audit}
\alias{gisco_get_urban_audit}
\title{Urban Audit dataset}
\source{
\url{https://gisco-services.ec.europa.eu/distribution/v2/}.

Copyright:
\url{https://ec.europa.eu/eurostat/web/gisco/geodata/administrative-units}.
}
\usage{
gisco_get_urban_audit(
  year = 2024,
  epsg = 4326,
  cache = TRUE,
  update_cache = FALSE,
  cache_dir = NULL,
  verbose = FALSE,
  spatialtype = c("RG", "LB"),
  country = NULL,
  level = c("all", "CITIES", "FUA", "GREATER_CITIES", "CITY", "KERN", "LUZ"),
  ext = "gpkg"
)
}
\arguments{
\item{year}{A character string or numeric value with the release year of the
file. One of
\Sexpr[stage=render,results=rd]{giscoR:::db_values("urban_audit",
"year",TRUE)}.}

\item{epsg}{A character string or numeric value with the coordinate
reference system as a 4-digit \href{https://epsg.io/}{EPSG code}. One of:
\itemize{
\item \code{"4326"}: \href{https://epsg.io/4326}{WGS84}.
\item \code{"3035"}: \href{https://epsg.io/3035}{ETRS89 / ETRS-LAEA}.
\item \code{"3857"}: \href{https://epsg.io/3857}{Pseudo-Mercator}.
}}

\item{cache}{A logical value indicating whether to cache results. Defaults
to \code{TRUE}. See \strong{Caching strategies} section in \code{\link[=gisco_set_cache_dir]{gisco_set_cache_dir()}}.}

\item{update_cache}{A logical value indicating whether to refresh the
cached file. Defaults to \code{FALSE}. When set to \code{TRUE}, it forces a new
download.}

\item{cache_dir}{A character string with a path to a cache directory. See
\strong{Caching strategies} section in \code{\link[=gisco_set_cache_dir]{gisco_set_cache_dir()}}.}

\item{verbose}{A logical value. If \code{TRUE} displays informational messages.}

\item{spatialtype}{A character string with the type of geometry to return.
Options available are:
\itemize{
\item \code{"RG"}: Regions - \code{MULTIPOLYGON/POLYGON} object.
\item \code{"LB"}: Labels - \code{POINT} object.
}}

\item{country}{A character vector of country codes. It can be either a
vector of country names, a vector of ISO 3166-1 alpha-3 country codes or a
vector of Eurostat country codes. See also \code{\link[countrycode:countrycode]{countrycode::countrycode()}}.}

\item{level}{A character string with the Urban Audit level. Possible values
are \code{"all"} (the default), which downloads the full dataset, \code{"CITIES"},
\code{"FUA"} and, for versions prior to \code{year = 2020}, \code{"GREATER_CITIES"},
\code{"CITY"}, \code{"KERN"} or \code{"LUZ"}.}

\item{ext}{A character value with the extension of the file (default
\code{"gpkg"}). One of
\Sexpr[stage=render,results=rd]{giscoR:::db_values("urban_audit",
"ext",TRUE)}.}
}
\value{
A \code{\link[sf:st_sf]{sf}} object.
}
\description{
This dataset contains the boundaries of cities (\code{"CITIES"}), greater cities
(\code{"GREATER_CITIES"}) and functional urban areas (\code{"FUA"}) defined according
to the EC-OECD city definition. It is used for the Eurostat Urban Audit data
collection.

Downloads data from the aggregated GISCO Urban Audit file. To download
single-unit Urban Audit files, use \code{\link[=gisco_get_unit_urban_audit]{gisco_get_unit_urban_audit()}}.
}
\details{
For more information, see: \href{https://ec.europa.eu/eurostat/statistics-explained/index.php?title=Territorial_typologies_for_European_cities_and_metropolitan_regions}{Eurostat - Statistics Explained}.

Cities are defined at several conceptual levels:
\itemize{
\item The core city (\code{"CITIES"}), using an administrative definition.
\item The Functional Urban Area/Large Urban Zone (\code{"FUA"}), approximating the
functional urban region.
Coverage includes the EU, Iceland, Norway and Switzerland. The dataset
includes polygon features, point features and a related attribute table
which can be joined on the URAU code field.
}

The \code{"URAU_CATG"} field defines the Urban Audit category:
\itemize{
\item \code{"C"} = City.
\item \code{"F"} = Functional urban area service type.
}
}
\section{Note}{
Check the download and usage provisions in \code{\link[=gisco_attributions]{gisco_attributions()}}.
}

\examples{
\dontshow{if (gisco_check_access()) withAutoprint(\{ # examplesIf}
\donttest{

cities <- gisco_get_urban_audit(year = 2024, level = "CITIES")

if (!is.null(cities)) {
  bcn <- cities[cities$URAU_NAME == "Barcelona", ]

  library(ggplot2)
  ggplot(bcn) +
    geom_sf()
}
}
\dontshow{\}) # examplesIf}
}
\seealso{
See \code{\link[=gisco_bulk_download]{gisco_bulk_download()}} to perform a bulk download of datasets.

See \code{\link[=gisco_get_unit_urban_audit]{gisco_get_unit_urban_audit()}} to download single-unit files.

Statistical unit datasets:
\code{\link[=gisco_get_census]{gisco_get_census()}},
\code{\link[=gisco_get_coastal_lines]{gisco_get_coastal_lines()}},
\code{\link[=gisco_get_lau]{gisco_get_lau()}},
\code{\link[=gisco_get_nuts]{gisco_get_nuts()}}
}
\concept{stats}
