% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gisco-get-urban-audit.R
\name{gisco_get_urban_audit}
\alias{gisco_get_urban_audit}
\title{Get GISCO greater cities and metropolitan areas \code{\link[sf:sf]{sf}} objects}
\source{
\url{https://gisco-services.ec.europa.eu/distribution/v2/}

Copyright:
\url{https://ec.europa.eu/eurostat/web/gisco/geodata/administrative-units}
}
\usage{
gisco_get_urban_audit(
  year = "2021",
  epsg = "4326",
  cache = TRUE,
  update_cache = FALSE,
  cache_dir = NULL,
  verbose = FALSE,
  spatialtype = "RG",
  country = NULL,
  level = NULL
)
}
\arguments{
\item{year}{Release year of the file. One of
\Sexpr[stage=render,results=rd]{giscoR:::for_docs("urban_audit",
"year",TRUE)}`.}

\item{epsg}{character string or number. Projection of the map: 4-digit
\href{https://epsg.io/}{EPSG code}. One of:
\itemize{
\item \code{"4326"}: \href{https://epsg.io/4326}{WGS84}
\item \code{"3035"}: \href{https://epsg.io/3035}{ETRS89 / ETRS-LAEA}
\item \code{"3857"}: \href{https://epsg.io/3857}{Pseudo-Mercator}
}}

\item{cache}{logical. Whether to do caching. Default is \code{TRUE}. See
\strong{Caching strategies} section in \code{\link[=gisco_set_cache_dir]{gisco_set_cache_dir()}}.}

\item{update_cache}{.logical. Should the cached file be refreshed?. Default
is \code{FALSE}. When set to \code{TRUE} it would force a new download.}

\item{cache_dir}{character string. A path to a cache directory. See
\strong{Caching strategies} section in \code{\link[=gisco_set_cache_dir]{gisco_set_cache_dir()}}.}

\item{verbose}{logical. If \code{TRUE} displays informational messages.}

\item{spatialtype}{Type of geometry to be returned:
\itemize{
\item \code{"LB"}: Labels - \code{POINT} object.
\item \code{"RG"}: Regions - \code{MULTIPOLYGON/POLYGON} object.
}}

\item{country}{Optional. A character vector of country codes. It could be
either a vector of country names, a vector of ISO3 country codes or a
vector of Eurostat country codes. See also \code{\link[countrycode:countrycode]{countrycode::countrycode()}}.}

\item{level}{Level of Urban Audit. Possible values are \code{"CITIES"}, \code{"FUA"},
\code{"GREATER_CITIES"} or \code{NULL}, that would download the full dataset.}
}
\value{
A \code{\link[sf:sf]{sf}} object specified by \code{spatialtype}.
}
\description{
Returns polygons and points corresponding to cities, greater cities and
metropolitan areas included on the
\href{https://ec.europa.eu/eurostat/web/regions-and-cities}{Urban Audit report}
of Eurostat.
}
\note{
Please check the download and usage provisions on \code{\link[=gisco_attributions]{gisco_attributions()}}.
}
\examples{
\dontshow{if (gisco_check_access()) withAutoprint(\{ # examplesIf}
\donttest{
cities <- gisco_get_urban_audit(year = "2020", level = "CITIES")

if (!is.null(cities)) {
  bcn <- cities[cities$URAU_NAME == "Barcelona", ]

  library(ggplot2)
  ggplot(bcn) +
    geom_sf()
}
}
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=gisco_get_communes]{gisco_get_communes()}}, \code{\link[=gisco_get_lau]{gisco_get_lau()}}

Other political: 
\code{\link{gisco_bulk_download}()},
\code{\link{gisco_get_units}()}
}
\concept{political}
