% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gisco_get_urban_audit.R
\name{gisco_get_urban_audit}
\alias{gisco_get_urban_audit}
\title{Download Geospatial Urban Audit Data from GISCO}
\source{
\href{https://gisco-services.ec.europa.eu/distribution/v2/urau/}{GISCO Urban Audit}
}
\usage{
gisco_get_urban_audit(
  year = "2018",
  epsg = "4326",
  update_cache = FALSE,
  cache_dir = NULL,
  spatialtype = "RG",
  level = NULL,
  country = NULL
)
}
\arguments{
\item{year}{Release year. One of
"2014", "2018" or "2020"}

\item{epsg}{projection of the map: 4-digit \href{https://spatialreference.org/ref/epsg/}{EPSG code}. One of:
\itemize{
\item "4326" - WGS84
\item "3035" - ETRS89 / ETRS-LAEA
\item "3857" - Pseudo-Mercator
}}

\item{update_cache}{a logical whether to update cache.}

\item{cache_dir}{a path to a cache directory. The directory have to exist.  The \code{NULL} (default) uses and creates \code{/gisco} directory in the temporary directory from \code{\link{tempdir}}. The directory can also be set with \code{options(gisco_cache_dir = <path>)}.}

\item{spatialtype}{Type of geometry to be returned:
\itemize{
 \item RG: Regions - Multipolygon
 \item LB: Labels - Point
}}

\item{level}{Level of Urban Audit. Possible values are 'CITIES', 'FUA', 'GREATER_CITIES' or \code{NULL}. See Details.}

\item{country}{Optional. A character vector of ISO-3 country codes.}
}
\value{
a \code{sf} object.
}
\description{
Downloads a simple feature (\code{sf}) object.
}
\details{
\code{level = NULL} would download the whole dataset including all levels

You can convert Eurostat country codes to ISO3 codes using the \code{\link[countrycode]{countrycode}} function:

eurostat_codes <- c("ES","UK","EL","PL","PT")\cr
\cr
countrycode::countrycode(\cr
  eurostat_codes,\cr
  origin = "eurostat",\cr
  destination = "iso3c"\cr
)
}
\note{
Please check the download and usage provisions on \link{gisco_attributions}.
}
\examples{
\donttest{
library(sf)
library(cartography)
europe <-
  gisco_get_countries(
    epsg = 3857,
    year = "2020",
    region = "Europe",
    resolution = "03"
  )
cities <-
  gisco_get_urban_audit(
    year = 2020,
    epsg = 3857,
    level = "GREATER_CITIES",
    country = "BEL"
  )

# Focus on Belgium
bbox <-
  st_bbox(c(
    xmin = 150000,
    xmax = 950000,
    ymax = 6900000,
    ymin = 6300000
  ),
  crs = st_crs(europe))
bbox <- st_bbox(cities)

# Plot
opar <- par(no.readonly = TRUE)
par(mar = c(1, 1, 1, 1))
plot(
  st_geometry(europe),
  xlim = bbox[c(1, 3)],
  ylim = bbox[c(2, 4)],
  col = "antiquewhite",
  graticule = TRUE
)
box()
plot(st_geometry(cities),
     col = "darkblue",
     border = "white",
     add = TRUE)

# Labels
labelLayer(
  st_crop(europe, bbox),
  txt = "NAME_ENGL",
  family = "serif",
  font = 3,
  cex = 0.8
)
labelLayer(
  cities,
  txt = "URAU_NAME",
  overlap = FALSE,
  col = "darkblue",
  halo = TRUE
)
layoutLayer(
  "Greater Cities of Belgium - Eurostat (2020)",
  col = "darkblue",
  sources = gisco_attributions(copyright = FALSE),
  horiz = FALSE,
  posscale = "bottomleft"
)
par(opar)
}
}
\author{
dieghernan, \url{https://github.com/dieghernan/}
}
