% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gisco_data.R
\docType{data}
\encoding{UTF-8}
\name{gisco_nuts}
\alias{gisco_nuts}
\title{All NUTS \code{POLYGON} object}
\format{
A \code{POLYGON} data frame (resolution: 1:20million, EPSG:4326) object with
11 variables:
\itemize{
\item \strong{id}: row ID
\item \strong{COAST_TYPE}: COAST_TYPE
\item \strong{MOUNT_TYPE}: MOUNT_TYPE
\item \strong{NAME_LATN}: Name on Latin characters
\item \strong{CNTR_CODE}: Eurostat Country code
\item \strong{FID}: FID
\item \strong{NUTS_ID}: NUTS identifier
\item \strong{NUTS_NAME}: NUTS name on local alphabet
\item \strong{LEVL_CODE}: NUTS level code (0,1,2,3)
\item \strong{URBN_TYPE}: URBN_TYPE
\item \strong{geometry}: geometry field
}
}
\source{
\url{https://gisco-services.ec.europa.eu/distribution/v2/nuts/geojson/}
\code{NUTS_RG_20M_2016_4326.geojson} file.
}
\description{
A \code{sf} object including all NUTS levels as provided by GISCO (2016 version).
}
\examples{

library(sf)

nuts <- gisco_nuts

italy <- subset(nuts, CNTR_CODE == "IT" & LEVL_CODE == 3)

plot(st_geometry(italy), col = c("springgreen4", "ivory", "red2"))
title(
  sub = gisco_attributions(),
  line = 1,
  cex.sub = 0.7,
  font.sub = 3
)
}
\seealso{
\code{\link[=gisco_get_nuts]{gisco_get_nuts()}}
}
\concept{dataset}
