% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giscoR-package.R
\docType{package}
\name{giscoR-package}
\alias{giscoR}
\alias{giscoR-package}
\title{giscoR: Download 'Eurostat' 'GISCO' Spatial Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools to download global and European spatial data from the 'Eurostat' 'GISCO' (Geographic Information System of the Commission) data distribution \url{https://ec.europa.eu/eurostat/web/gisco}. The package provides helpers for country boundaries, 'NUTS' regions, administrative units, statistical units, transport networks, basic service locations and other 'GISCO' datasets. This package is not officially related to or endorsed by 'Eurostat'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ropengov.github.io/giscoR/}
  \item \url{https://github.com/rOpenGov/giscoR}
  \item Report bugs at \url{https://github.com/rOpenGov/giscoR/issues}
}

}
\author{
\strong{Maintainer}: Diego Hernangómez \email{diego.hernangomezherrero@gmail.com} (\href{https://orcid.org/0000-0001-8457-4658}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Diego Hernangómez \email{diego.hernangomezherrero@gmail.com} (\href{https://orcid.org/0000-0001-8457-4658}{ORCID}) [copyright holder]
}

Other contributors:
\itemize{
  \item Eurostat (\href{https://ror.org/033d3q980}{ROR}) [copyright holder]
  \item EuroGeographics [copyright holder]
}

}
\keyword{internal}
