% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giscoR-package.R
\docType{package}
\name{giscoR-package}
\alias{giscoR-package}
\alias{giscoR}
\title{Download geospatial data from GISCO API - Eurostat}
\source{
\url{https://ec.europa.eu/eurostat/web/gisco/geodata/reference-data/administrative-units-statistical-units}
}
\description{
giscoR is a API package that helps to retrieve data from Eurostat - GISCO (the Geographic Information System of the COmmission)
}
\details{
giscoR package

\tabular{ll}{
Package: \tab giscoR \cr
Type: \tab Package\cr
Version: \tab See sessionInfo() or DESCRIPTION file\cr
Date: \tab 2020\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}
}
\examples{
library(giscoR)
}
\references{
See \code{citation("giscoR")}
}
\author{
dieghernan, \url{https://github.com/dieghernan/}
}
\keyword{package}
