% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giscoR-package.R
\docType{package}
\name{giscoR-package}
\alias{giscoR}
\alias{giscoR-package}
\title{giscoR: Download Map Data from GISCO API - Eurostat}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools to download data from the GISCO (Geographic Information System of the Commission) Eurostat database \url{https://ec.europa.eu/eurostat/web/gisco}. Global and European map data available. This package is in no way officially related to or endorsed by Eurostat.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ropengov.github.io/giscoR/}
  \item \url{https://github.com/rOpenGov/giscoR}
  \item Report bugs at \url{https://github.com/rOpenGov/giscoR/issues}
}

}
\author{
\strong{Maintainer}: Diego Hernangómez \email{diego.hernangomezherrero@gmail.com} (\href{https://orcid.org/0000-0001-8457-4658}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item EuroGeographics (for the administrative boundaries.) [copyright holder]
  \item Vincent Arel-Bundock (\href{https://orcid.org/0000-0003-2042-7063}{ORCID}) (for the gisco_countrycode dataset.) [copyright holder]
}

}
\keyword{internal}
