% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gisco_data.R
\docType{data}
\name{tgs00026}
\alias{tgs00026}
\title{Disposable income of private households by NUTS 2 regions}
\format{
data_frame:
\itemize{
\item \strong{geo}: NUTS2 identifier
\item \strong{time}: reference year (2007 to 2018)
\item \strong{values}: value in euros
}
}
\source{
\url{https://ec.europa.eu/eurostat}, extracted on 2020-10-27
}
\description{
The disposable income of private households is the balance
of primary income (operating surplus/mixed income plus compensation of
employees plus property income received minus property income paid) and
the redistribution of income in cash. These transactions comprise social
contributions paid, social benefits in cash received, current taxes on
income and wealth paid, as well as other current transfers. Disposable
income does not include social transfers in kind coming from public
administrations or non-profit institutions serving households.
}
\examples{

data(tgs00026)
}
\concept{dataset}
